fix(i18n): wire Vietnamese locale (#774)
