+++ packages/core/src/i18n/i18n-translate.ts
@@ -594,6 +594,60 @@ const nlPairs: Array<[string, string]> = [
         ['days', 'dagen'],
     ];
 
+const viPairs: Array<[string, string]> = [
+        ['Settings', 'Cài đặt'],
+        ['General', 'Chung'],
+        ['Appearance', 'Giao diện'],
+        ['Language', 'Ngôn ngữ'],
+        ['Projects', 'Dự án'],
+        ['Project', 'Dự án'],
+        ['Tasks', 'Nhiệm vụ'],
+        ['Task', 'Nhiệm vụ'],
+        ['Inbox', 'Hộp thư đến'],
+        ['Next Actions', 'Hành động tiếp theo'],
+        ['Next', 'Tiếp theo'],
+        ['Waiting', 'Đang chờ'],
+        ['Someday/Maybe', 'Tương lai/Có thể'],
+        ['Calendar', 'Lịch'],
+        ['Review', 'Xem lại'],
+        ['Archive', 'Lưu trữ'],
+        ['Archived', 'Đã lưu trữ'],
+        ['Trash', 'Thùng rác'],
+        ['Add', 'Thêm'],
+        ['Delete', 'Xóa'],
+        ['Save', 'Lưu'],
+        ['Cancel', 'Hủy'],
+        ['Search', 'Tìm kiếm'],
+        ['Notifications', 'Thông báo'],
+        ['Sync', 'Đồng bộ'],
+        ['Data', 'Dữ liệu'],
+        ['About', 'Giới thiệu'],
+        ['Advanced', 'Nâng cao'],
+        ['Done', 'Hoàn thành'],
+        ['Completed', 'Đã hoàn thành'],
+        ['Success', 'Thành công'],
+        ['Error', 'Lỗi'],
+        ['Notice', 'Thông báo'],
+        ['Update Available', 'Có bản cập nhật'],
+        ['Up to Date', 'Đã cập nhật'],
+        ['Changelog', 'Nhật ký thay đổi'],
+        ['Download', 'Tải xuống'],
+        ['Later', 'Để sau'],
+        ['Loading', 'Đang tải'],
+        ['Month', 'Tháng'],
+        ['Backup', 'Sao lưu'],
+        ['License', 'Giấy phép'],
+        ['Website', 'Trang web'],
+        ['Optional', 'Tùy chọn'],
+        ['Test', 'Kiểm tra'],
+        ['Last Sync', 'Lần đồng bộ cuối'],
+        ['Never', 'Không bao giờ'],
+        ['Conflicts', 'Xung đột'],
+        ['No free time', 'Không còn thời gian trống'],
+        ['Time conflict', 'Trùng thời gian'],
+        ['days', 'ngày'],
+];
+
 const translationPairs: Record<Language, Array<[string, string]>> = {
         en: [],
         zh: [],
@@ -768,6 +822,7 @@ const translationPairs: Record<Language, Array<[string, string]>> = {
         tr: trPairs,
         nl: nlPairs,
         cs: csPairs,
+        vi: viPairs,
 };
 
 const escapeRegExp = (value: string) => value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
