+++ packages/core/src/i18n/i18n-translations.ts
@@ -17,6 +17,7 @@ import { itOverrides } from './locales/it';
 import { trOverrides } from './locales/tr';
 import { nlOverrides } from './locales/nl';
 import { csOverrides } from './locales/cs';
+import { viOverrides } from './locales/vi';
 
 const buildTranslations = (overrides: Record<string, string>) => {
     const result: Record<string, string> = {};
@@ -40,6 +41,7 @@ const it = buildTranslations(itOverrides);
 const tr = buildTranslations(trOverrides);
 const nl = buildTranslations(nlOverrides);
 const cs = buildTranslations(csOverrides);
+const vi = buildTranslations(viOverrides);
 
 export const translations: Record<Language, Record<string, string>> = {
     en,
@@ -59,4 +61,5 @@ export const translations: Record<Language, Record<string, string>> = {
     tr,
     nl,
     cs,
+    vi,
 };
