--- packages/core/src/i18n/locales/i18n-translate.ts
@@ -1,877 +1,0 @@
-import type { Language } from './i18n-types';
-const csPairs: Array<[string, string]> = [
-        ['Settings', 'Nastavení'],
-        ['General', 'Obecné'],
-        ['Appearance', 'Vzhled'],
-        ['Language', 'Jazyk'],
-        ['Projects', 'Projekty'],
-        ['Project', 'Projekt'],
-        ['Tasks', 'Úkoly'],
-        ['Task', 'Úkol'],
-        ['Inbox', 'Schránka'],
-        ['Next Actions', 'Další kroky'],
-        ['Next', 'Další'],
-        ['Waiting', 'Čeká'],
-        ['Someday/Maybe', 'Někdy/Možná'],
-        ['Calendar', 'Kalendář'],
-        ['Review', 'Revize'],
-        ['Archive', 'Archiv'],
-        ['Archived', 'Archivováno'],
-        ['Trash', 'Koš'],
-        ['Add', 'Přidat'],
-        ['Delete', 'Smazat'],
-        ['Save', 'Uložit'],
-        ['Cancel', 'Zrušit'],
-        ['Search', 'Hledat'],
-        ['Notifications', 'Oznámení'],
-        ['Sync', 'Synchronizace'],
-        ['Data', 'Data'],
-        ['About', 'O aplikaci'],
-        ['Advanced', 'Pokročilé'],
-        ['Done', 'Hotovo'],
-        ['Completed', 'Dokončeno'],
-        ['Success', 'Úspěch'],
-        ['Error', 'Chyba'],
-        ['Notice', 'Upozornění'],
-        ['Update Available', 'Aktualizace dostupná'],
-        ['Up to Date', 'Aktuální'],
-        ['Changelog', 'Seznam změn'],
-        ['Download', 'Stáhnout'],
-        ['Later', 'Později'],
-        ['Loading', 'Načítání'],
-        ['Month', 'Měsíc'],
-        ['Backup', 'Záloha'],
-        ['License', 'Licence'],
-        ['Website', 'Web'],
-        ['Optional', 'Volitelné'],
-        ['Test', 'Test'],
-        ['Last Sync', 'Poslední synchronizace'],
-        ['Never', 'Nikdy'],
-        ['Conflicts', 'Konflikty'],
-        ['No free time', 'Žádný volný čas'],
-        ['Time conflict', 'Časový konflikt'],
-        ['days', 'dní'],
-    ];
-
-const dePairs: Array<[string, string]> = [
-        ['Settings', 'Einstellungen'],
-        ['General', 'Allgemein'],
-        ['Appearance', 'Darstellung'],
-        ['Language', 'Sprache'],
-        ['Projects', 'Projekte'],
-        ['Project', 'Projekt'],
-        ['Tasks', 'Aufgaben'],
-        ['Task', 'Aufgabe'],
-        ['Inbox', 'Posteingang'],
-        ['Next Actions', 'Nächste Aktionen'],
-        ['Next', 'Nächste'],
-        ['Waiting', 'Warten'],
-        ['Someday/Maybe', 'Irgendwann/Vielleicht'],
-        ['Calendar', 'Kalender'],
-        ['Review', 'Review'],
-        ['Settings', 'Einstellungen'],
-        ['Archive', 'Archiv'],
-        ['Archived', 'Archiviert'],
-        ['Trash', 'Papierkorb'],
-        ['Add', 'Hinzufügen'],
-        ['Delete', 'Löschen'],
-        ['Save', 'Speichern'],
-        ['Cancel', 'Abbrechen'],
-        ['Search', 'Suchen'],
-        ['Notifications', 'Benachrichtigungen'],
-        ['Sync', 'Sync'],
-        ['Data', 'Daten'],
-        ['About', 'Über'],
-        ['Advanced', 'Erweitert'],
-        ['Done', 'Erledigt'],
-        ['Completed', 'Abgeschlossen'],
-        ['Success', 'Erfolg'],
-        ['Error', 'Fehler'],
-        ['Notice', 'Hinweis'],
-        ['Update Available', 'Update verfügbar'],
-        ['Up to Date', 'Aktuell'],
-        ['Changelog', 'Änderungsprotokoll'],
-        ['Download', 'Download'],
-        ['Later', 'Später'],
-        ['Loading', 'Laden'],
-        ['Month', 'Monat'],
-        ['Backup', 'Sicherung'],
-        ['License', 'Lizenz'],
-        ['Website', 'Webseite'],
-        ['Optional', 'Optional'],
-        ['Test', 'Test'],
-        ['Last Sync', 'Letzte Synchronisierung'],
-        ['Never', 'Nie'],
-        ['Conflicts', 'Konflikte'],
-        ['No free time', 'Keine freie Zeit'],
-        ['Time conflict', 'Zeitkonflikt'],
-        ['days', 'Tage'],
-    ];
-
-const frPairs: Array<[string, string]> = [
-        ['Settings', 'Paramètres'],
-        ['General', 'Général'],
-        ['Appearance', 'Apparence'],
-        ['Language', 'Langue'],
-        ['Projects', 'Projets'],
-        ['Project', 'Projet'],
-        ['Tasks', 'Tâches'],
-        ['Task', 'Tâche'],
-        ['Inbox', 'Boîte de réception'],
-        ['Next Actions', 'Prochaines actions'],
-        ['Next', 'Suivant'],
-        ['Waiting', 'En attente'],
-        ['Someday/Maybe', 'Un jour/Peut‑être'],
-        ['Calendar', 'Calendrier'],
-        ['Review', 'Revue'],
-        ['Archive', 'Archive'],
-        ['Archived', 'Archivé'],
-        ['Trash', 'Corbeille'],
-        ['Add', 'Ajouter'],
-        ['Delete', 'Supprimer'],
-        ['Save', 'Enregistrer'],
-        ['Cancel', 'Annuler'],
-        ['Search', 'Rechercher'],
-        ['Notifications', 'Notifications'],
-        ['Sync', 'Synchronisation'],
-        ['Data', 'Données'],
-        ['About', 'À propos'],
-        ['Advanced', 'Avancé'],
-        ['Done', 'Terminé'],
-        ['Completed', 'Terminé'],
-        ['Success', 'Succès'],
-        ['Error', 'Erreur'],
-        ['Notice', 'Avis'],
-        ['Update Available', 'Mise à jour disponible'],
-        ['Up to Date', 'À jour'],
-        ['Changelog', 'Journal des modifications'],
-        ['Download', 'Télécharger'],
-        ['Later', 'Plus tard'],
-        ['Loading', 'Chargement'],
-        ['Month', 'Mois'],
-        ['Backup', 'Sauvegarde'],
-        ['License', 'Licence'],
-        ['Website', 'Site web'],
-        ['Optional', 'Facultatif'],
-        ['Test', 'Tester'],
-        ['Last Sync', 'Dernière synchronisation'],
-        ['Never', 'Jamais'],
-        ['Conflicts', 'Conflits'],
-        ['No free time', 'Pas de temps libre'],
-        ['Time conflict', 'Conflit de temps'],
-        ['days', 'jours'],
-    ];
-
-const ptPairs: Array<[string, string]> = [
-        ['Settings', 'Configurações'],
-        ['General', 'Geral'],
-        ['Appearance', 'Aparência'],
-        ['Language', 'Idioma'],
-        ['Projects', 'Projetos'],
-        ['Project', 'Projeto'],
-        ['Tasks', 'Tarefas'],
-        ['Task', 'Tarefa'],
-        ['Inbox', 'Caixa de entrada'],
-        ['Next Actions', 'Próximas ações'],
-        ['Next', 'Próximo'],
-        ['Waiting', 'Aguardando'],
-        ['Someday/Maybe', 'Algum dia/Talvez'],
-        ['Calendar', 'Calendário'],
-        ['Review', 'Revisão'],
-        ['Archive', 'Arquivo'],
-        ['Archived', 'Arquivado'],
-        ['Trash', 'Lixeira'],
-        ['Add', 'Adicionar'],
-        ['Delete', 'Excluir'],
-        ['Save', 'Salvar'],
-        ['Cancel', 'Cancelar'],
-        ['Search', 'Pesquisar'],
-        ['Notifications', 'Notificações'],
-        ['Sync', 'Sincronização'],
-        ['Data', 'Dados'],
-        ['About', 'Sobre'],
-        ['Advanced', 'Avançado'],
-        ['Done', 'Concluído'],
-        ['Completed', 'Concluído'],
-        ['Success', 'Sucesso'],
-        ['Error', 'Erro'],
-        ['Notice', 'Aviso'],
-        ['Update Available', 'Atualização disponível'],
-        ['Up to Date', 'Atualizado'],
-        ['Changelog', 'Registro de alterações'],
-        ['Download', 'Baixar'],
-        ['Later', 'Mais tarde'],
-        ['Loading', 'Carregando'],
-        ['Month', 'Mês'],
-        ['Backup', 'Backup'],
-        ['License', 'Licença'],
-        ['Website', 'Site'],
-        ['Optional', 'Opcional'],
-        ['Test', 'Testar'],
-        ['Last Sync', 'Última sincronização'],
-        ['Never', 'Nunca'],
-        ['Conflicts', 'Conflitos'],
-        ['No free time', 'Sem tempo livre'],
-        ['Time conflict', 'Conflito de tempo'],
-        ['days', 'dias'],
-    ];
-
-const itPairs: Array<[string, string]> = [
-        ['Settings', 'Impostazioni'],
-        ['General', 'Generale'],
-        ['Appearance', 'Aspetto'],
-        ['Language', 'Lingua'],
-        ['Projects', 'Progetti'],
-        ['Project', 'Progetto'],
-        ['Tasks', 'Attività'],
-        ['Task', 'Attività'],
-        ['Inbox', 'Entrata'],
-        ['Next Actions', 'Prossime azioni'],
-        ['Next', 'Prossimo'],
-        ['Waiting', 'In attesa'],
-        ['Someday/Maybe', 'Un giorno/Forse'],
-        ['Calendar', 'Calendario'],
-        ['Review', 'Revisione'],
-        ['Archive', 'Archivio'],
-        ['Archived', 'Archiviato'],
-        ['Trash', 'Cestino'],
-        ['Add', 'Aggiungi'],
-        ['Delete', 'Elimina'],
-        ['Save', 'Salva'],
-        ['Cancel', 'Annulla'],
-        ['Search', 'Cerca'],
-        ['Notifications', 'Notifiche'],
-        ['Sync', 'Sincronizzazione'],
-        ['Data', 'Dati'],
-        ['About', 'Informazioni'],
-        ['Advanced', 'Avanzate'],
-        ['Done', 'Fatto'],
-        ['Completed', 'Completato'],
-        ['Success', 'Successo'],
-        ['Error', 'Errore'],
-        ['Notice', 'Avviso'],
-        ['Update Available', 'Aggiornamento disponibile'],
-        ['Up to Date', 'Aggiornato'],
-        ['Changelog', 'Changelog'],
-        ['Download', 'Scarica'],
-        ['Later', 'Più tardi'],
-        ['Loading', 'Caricamento'],
-        ['Month', 'Mese'],
-        ['Backup', 'Backup'],
-        ['License', 'Licenza'],
-        ['Website', 'Sito web'],
-        ['Optional', 'Opzionale'],
-        ['Test', 'Test'],
-        ['Last Sync', 'Ultima sinc.'],
-        ['Never', 'Mai'],
-        ['Conflicts', 'Conflitti'],
-        ['No free time', 'Nessuno slot libero'],
-        ['Time conflict', 'Conflitto di orario'],
-        ['days', 'giorni'],
-    ];
-
-const trPairs: Array<[string, string]> = [
-        ['Settings', 'Ayarlar'],
-        ['General', 'Genel'],
-        ['Appearance', 'Görünüm'],
-        ['Language', 'Dil'],
-        ['Projects', 'Projeler'],
-        ['Project', 'Proje'],
-        ['Tasks', 'Görevler'],
-        ['Task', 'Görev'],
-        ['Inbox', 'Gelen Kutusu'],
-        ['Next Actions', 'Sonraki adımlar'],
-        ['Next', 'Sonraki'],
-        ['Waiting', 'Beklemede'],
-        ['Someday/Maybe', 'Bir gün/Belki'],
-        ['Calendar', 'Takvim'],
-        ['Review', 'Gözden geçirme'],
-        ['Archive', 'Arşiv'],
-        ['Archived', 'Arşivlendi'],
-        ['Trash', 'Çöp kutusu'],
-        ['Add', 'Ekle'],
-        ['Delete', 'Sil'],
-        ['Save', 'Kaydet'],
-        ['Cancel', 'İptal'],
-        ['Search', 'Ara'],
-        ['Notifications', 'Bildirimler'],
-        ['Sync', 'Eşitleme'],
-        ['Data', 'Veri'],
-        ['About', 'Hakkında'],
-        ['Advanced', 'Gelişmiş'],
-        ['Done', 'Tamam'],
-        ['Completed', 'Tamamlandı'],
-        ['Success', 'Başarılı'],
-        ['Error', 'Hata'],
-        ['Notice', 'Bildirim'],
-        ['Update Available', 'Güncelleme mevcut'],
-        ['Up to Date', 'Güncel'],
-        ['Changelog', 'Değişiklik günlüğü'],
-        ['Download', 'İndir'],
-        ['Later', 'Daha sonra'],
-        ['Loading', 'Yükleniyor'],
-        ['Month', 'Ay'],
-        ['Backup', 'Yedek'],
-        ['License', 'Lisans'],
-        ['Website', 'Web sitesi'],
-        ['Optional', 'İsteğe bağlı'],
-        ['Test', 'Test'],
-        ['Last Sync', 'Son senkronizasyon'],
-        ['Never', 'Asla'],
-        ['Conflicts', 'Çakışmalar'],
-        ['No free time', 'Boş zaman yok'],
-        ['Time conflict', 'Zaman çakışması'],
-        ['days', 'gün'],
-    ];
-
-const ruPairs: Array<[string, string]> = [
-        ['Settings', 'Настройки'],
-        ['General', 'Общее'],
-        ['Appearance', 'Внешний вид'],
-        ['Language', 'Язык'],
-        ['Projects', 'Проекты'],
-        ['Project', 'Проект'],
-        ['Tasks', 'Задачи'],
-        ['Task', 'Задача'],
-        ['Inbox', 'Входящие'],
-        ['Next Actions', 'Следующие действия'],
-        ['Next', 'Следующее'],
-        ['Waiting', 'Ожидание'],
-        ['Someday/Maybe', 'Когда-нибудь/Возможно'],
-        ['Calendar', 'Календарь'],
-        ['Review', 'Обзор'],
-        ['Archive', 'Архив'],
-        ['Archived', 'Архивировано'],
-        ['Trash', 'Корзина'],
-        ['Add', 'Добавить'],
-        ['Delete', 'Удалить'],
-        ['Save', 'Сохранить'],
-        ['Cancel', 'Отмена'],
-        ['Search', 'Поиск'],
-        ['Notifications', 'Уведомления'],
-        ['Sync', 'Синхронизация'],
-        ['Data', 'Данные'],
-        ['About', 'О приложении'],
-        ['Advanced', 'Дополнительно'],
-        ['Done', 'Готово'],
-        ['Completed', 'Завершено'],
-        ['Success', 'Успешно'],
-        ['Error', 'Ошибка'],
-        ['Notice', 'Уведомление'],
-        ['Update Available', 'Доступно обновление'],
-        ['Up to Date', 'Актуально'],
-        ['Changelog', 'Журнал изменений'],
-        ['Download', 'Скачать'],
-        ['Later', 'Позже'],
-        ['Loading', 'Загрузка'],
-        ['Month', 'Месяц'],
-        ['Backup', 'Резервная копия'],
-        ['License', 'Лицензия'],
-        ['Website', 'Сайт'],
-        ['Optional', 'Необязательно'],
-        ['Test', 'Тест'],
-        ['Last Sync', 'Последняя синхронизация'],
-        ['Never', 'Никогда'],
-        ['Conflicts', 'Конфликты'],
-        ['No free time', 'Нет свободного времени'],
-        ['Time conflict', 'Конфликт времени'],
-        ['days', 'дней'],
-    ];
-
-const jaPairs: Array<[string, string]> = [
-        ['Settings', '設定'],
-        ['General', '一般'],
-        ['Appearance', '外観'],
-        ['Language', '言語'],
-        ['Projects', 'プロジェクト'],
-        ['Project', 'プロジェクト'],
-        ['Tasks', 'タスク'],
-        ['Task', 'タスク'],
-        ['Inbox', '受信箱'],
-        ['Next Actions', '次のアクション'],
-        ['Next', '次へ'],
-        ['Waiting', '待機中'],
-        ['Someday/Maybe', 'いつか/たぶん'],
-        ['Calendar', 'カレンダー'],
-        ['Review', 'レビュー'],
-        ['Archive', 'アーカイブ'],
-        ['Archived', 'アーカイブ済み'],
-        ['Trash', 'ゴミ箱'],
-        ['Add', '追加'],
-        ['Delete', '削除'],
-        ['Save', '保存'],
-        ['Cancel', 'キャンセル'],
-        ['Search', '検索'],
-        ['Notifications', '通知'],
-        ['Sync', '同期'],
-        ['Data', 'データ'],
-        ['About', '情報'],
-        ['Advanced', '詳細'],
-        ['Done', '完了'],
-        ['Completed', '完了'],
-        ['Success', '成功'],
-        ['Error', 'エラー'],
-        ['Notice', '注意'],
-        ['Update Available', '更新があります'],
-        ['Up to Date', '最新'],
-        ['Changelog', '変更履歴'],
-        ['Download', 'ダウンロード'],
-        ['Later', '後で'],
-        ['Loading', '読み込み中'],
-        ['Month', '月'],
-        ['Backup', 'バックアップ'],
-        ['License', 'ライセンス'],
-        ['Website', 'ウェブサイト'],
-        ['Optional', '任意'],
-        ['Test', 'テスト'],
-        ['Last Sync', '最終同期'],
-        ['Never', 'なし'],
-        ['Conflicts', '競合'],
-        ['No free time', '空き時間なし'],
-        ['Time conflict', '時間の競合'],
-        ['days', '日'],
-    ];
-
-const koPairs: Array<[string, string]> = [
-        ['Settings', '설정'],
-        ['General', '일반'],
-        ['Appearance', '모양'],
-        ['Language', '언어'],
-        ['Projects', '프로젝트'],
-        ['Project', '프로젝트'],
-        ['Tasks', '작업'],
-        ['Task', '작업'],
-        ['Inbox', '받은편지함'],
-        ['Next Actions', '다음 작업'],
-        ['Next', '다음'],
-        ['Waiting', '대기'],
-        ['Someday/Maybe', '언젠가/아마도'],
-        ['Calendar', '캘린더'],
-        ['Review', '검토'],
-        ['Archive', '보관'],
-        ['Archived', '보관됨'],
-        ['Trash', '휴지통'],
-        ['Add', '추가'],
-        ['Delete', '삭제'],
-        ['Save', '저장'],
-        ['Cancel', '취소'],
-        ['Search', '검색'],
-        ['Notifications', '알림'],
-        ['Sync', '동기화'],
-        ['Data', '데이터'],
-        ['About', '정보'],
-        ['Advanced', '고급'],
-        ['Done', '완료'],
-        ['Completed', '완료'],
-        ['Success', '성공'],
-        ['Error', '오류'],
-        ['Notice', '알림'],
-        ['Update Available', '업데이트 가능'],
-        ['Up to Date', '최신'],
-        ['Changelog', '변경 로그'],
-        ['Download', '다운로드'],
-        ['Later', '나중에'],
-        ['Loading', '로딩 중'],
-        ['Month', '월'],
-        ['Backup', '백업'],
-        ['License', '라이선스'],
-        ['Website', '웹사이트'],
-        ['Optional', '선택 사항'],
-        ['Test', '테스트'],
-        ['Last Sync', '마지막 동기화'],
-        ['Never', '없음'],
-        ['Conflicts', '충돌'],
-        ['No free time', '빈 시간 없음'],
-        ['Time conflict', '시간 충돌'],
-        ['days', '일'],
-    ];
-
-const plPairs: Array<[string, string]> = [
-        ['Settings', 'Ustawienia'],
-        ['General', 'Ogólne'],
-        ['Appearance', 'Wygląd'],
-        ['Language', 'Język'],
-        ['Projects', 'Projekty'],
-        ['Project', 'Projekt'],
-        ['Tasks', 'Zadania'],
-        ['Task', 'Zadanie'],
-        ['Inbox', 'Skrzynka odbiorcza'],
-        ['Next Actions', 'Następne działania'],
-        ['Next', 'Następne'],
-        ['Waiting', 'Oczekujące'],
-        ['Someday/Maybe', 'Kiedyś/Może'],
-        ['Calendar', 'Kalendarz'],
-        ['Review', 'Przegląd'],
-        ['Archive', 'Archiwum'],
-        ['Archived', 'Zarchiwizowane'],
-        ['Trash', 'Kosz'],
-        ['Add', 'Dodaj'],
-        ['Delete', 'Usuń'],
-        ['Save', 'Zapisz'],
-        ['Cancel', 'Anuluj'],
-        ['Search', 'Szukaj'],
-        ['Notifications', 'Powiadomienia'],
-        ['Sync', 'Synchronizacja'],
-        ['Data', 'Dane'],
-        ['About', 'O aplikacji'],
-        ['Advanced', 'Zaawansowane'],
-        ['Done', 'Gotowe'],
-        ['Completed', 'Ukończone'],
-        ['Success', 'Sukces'],
-        ['Error', 'Błąd'],
-        ['Notice', 'Uwaga'],
-        ['Update Available', 'Dostępna aktualizacja'],
-        ['Up to Date', 'Aktualne'],
-        ['Changelog', 'Dziennik zmian'],
-        ['Download', 'Pobierz'],
-        ['Later', 'Później'],
-        ['Loading', 'Ładowanie'],
-        ['Month', 'Miesiąc'],
-        ['Backup', 'Kopia zapasowa'],
-        ['License', 'Licencja'],
-        ['Website', 'Strona internetowa'],
-        ['Optional', 'Opcjonalne'],
-        ['Test', 'Test'],
-        ['Last Sync', 'Ostatnia synchronizacja'],
-        ['Never', 'Nigdy'],
-        ['Conflicts', 'Konflikty'],
-        ['No free time', 'Brak wolnego czasu'],
-        ['Time conflict', 'Konflikt czasowy'],
-        ['days', 'dni'],
-    ];
-
-const nlPairs: Array<[string, string]> = [
-        ['Settings', 'Instellingen'],
-        ['General', 'Algemeen'],
-        ['Appearance', 'Uiterlijk'],
-        ['Language', 'Taal'],
-        ['Projects', 'Projecten'],
-        ['Project', 'Project'],
-        ['Tasks', 'Taken'],
-        ['Task', 'Taak'],
-        ['Inbox', 'Inbox'],
-        ['Next Actions', 'Volgende acties'],
-        ['Next', 'Volgende'],
-        ['Waiting', 'Wachten'],
-        ['Someday/Maybe', 'Ooit/Misschien'],
-        ['Calendar', 'Kalender'],
-        ['Review', 'Nalezen'],
-        ['Archive', 'Archief'],
-        ['Archived', 'Gearchiveerd'],
-        ['Trash', 'Prullenbak'],
-        ['Add', 'Toevoegen'],
-        ['Delete', 'Verwijderen'],
-        ['Save', 'Opslaan'],
-        ['Cancel', 'Annuleren'],
-        ['Search', 'Zoeken'],
-        ['Notifications', 'Meldingen'],
-        ['Sync', 'Synchronisatie'],
-        ['Data', 'Gegevens'],
-        ['About', 'Over'],
-        ['Advanced', 'Geavanceerd'],
-        ['Done', 'Klaar'],
-        ['Completed', 'Voltooid'],
-        ['Success', 'Succes'],
-        ['Error', 'Fout'],
-        ['Notice', 'Melding'],
-        ['Update Available', 'Update beschikbaar'],
-        ['Up to Date', 'Actueel'],
-        ['Changelog', 'Wijzigingslog'],
-        ['Download', 'Download'],
-        ['Later', 'Later'],
-        ['Loading', 'Laden'],
-        ['Month', 'Maand'],
-        ['Backup', 'Back-up'],
-        ['License', 'Licentie'],
-        ['Website', 'Website'],
-        ['Optional', 'Optioneel'],
-        ['Test', 'Test'],
-        ['Last Sync', 'Laatste synchronisatie'],
-        ['Never', 'Nooit'],
-        ['Conflicts', 'Conflicten'],
-        ['No free time', 'Geen vrije tijd'],
-        ['Time conflict', 'Tijdsconflict'],
-        ['days', 'dagen'],
-    ];
-const viPairs: Array<[string, string]> = [
-          ['Settings', 'Cài đặt'],
-          ['General', 'Chung'],
-          ['Appearance', 'Giao diện'],
-          ['Language', 'Ngôn ngữ'],
-          ['Projects', 'Dự án'],
-          ['Project', 'Dự án'],
-          ['Tasks', 'Nhiệm vụ'],
-          ['Task', 'Nhiệm vụ'],
-          ['Inbox', 'Hộp thư đến'],
-          ['Next Actions', 'Hành động tiếp theo'],
-          ['Next', 'Tiếp theo'],
-          ['Waiting', 'Đang chờ'],
-          ['Someday/Maybe', 'Tương lai/Có thể'],
-          ['Calendar', 'Lịch'],
-          ['Review', 'Xem lại'],
-          ['Archive', 'Lưu trữ'],
-          ['Archived', 'Đã lưu trữ'],
-          ['Trash', 'Thùng rác'],
-          ['Add', 'Thêm'],
-          ['Delete', 'Xóa'],
-          ['Save', 'Lưu'],
-          ['Cancel', 'Hủy'],
-          ['Search', 'Tìm kiếm'],
-          ['Notifications', 'Thông báo'],
-          ['Sync', 'Đồng bộ'],
-          ['Data', 'Dữ liệu'],
-          ['About', 'Giới thiệu'],
-          ['Advanced', 'Nâng cao'],
-          ['Done', 'Hoàn thành'],
-          ['Completed', 'Đã hoàn thành'],
-          ['Success', 'Thành công'],
-          ['Error', 'Lỗi'],
-          ['Notice', 'Thông báo'],
-          ['Update Available', 'Có bản cập nhật'],
-          ['Up to Date', 'Đã cập nhật'],
-          ['Changelog', 'Nhật ký thay đổi'],
-          ['Download', 'Tải xuống'],
-          ['Later', 'Để sau'],
-          ['Loading', 'Đang tải'],
-          ['Month', 'Tháng'],
-          ['Backup', 'Sao lưu'],
-          ['License', 'Giấy phép'],
-          ['Website', 'Trang web'],
-          ['Optional', 'Tùy chọn'],
-          ['Test', 'Kiểm tra'],
-          ['Last Sync', 'Lần đồng bộ cuối'],
-          ['Never', 'Không bao giờ'],
-          ['Conflicts', 'Xung đột'],
-          ['No free time', 'Không còn thời gian trống'],
-          ['Time conflict', 'Trùng thời gian'],
-          ['days', 'ngày'],
-];
-
-const translationPairs: Record<Language, Array<[string, string]>> = {
-        en: [],
-        zh: [],
-        'zh-Hant': [],
-        es: [
-            ['Settings', 'Ajustes'],
-            ['General', 'General'],
-            ['Appearance', 'Apariencia'],
-            ['Language', 'Idioma'],
-            ['Projects', 'Proyectos'],
-            ['Project', 'Proyecto'],
-            ['Tasks', 'Tareas'],
-            ['Task', 'Tarea'],
-            ['Inbox', 'Bandeja de entrada'],
-            ['Next Actions', 'Siguientes acciones'],
-            ['Next', 'Siguiente'],
-            ['Waiting', 'En espera'],
-            ['Someday/Maybe', 'Algún día/Tal vez'],
-            ['Calendar', 'Calendario'],
-            ['Review', 'Revisión'],
-            ['Archive', 'Archivo'],
-            ['Archived', 'Archivado'],
-            ['Trash', 'Papelera'],
-            ['Add', 'Añadir'],
-            ['Delete', 'Eliminar'],
-            ['Save', 'Guardar'],
-            ['Cancel', 'Cancelar'],
-            ['Search', 'Buscar'],
-            ['Notifications', 'Notificaciones'],
-            ['Sync', 'Sincronizar'],
-            ['Data', 'Datos'],
-            ['About', 'Acerca de'],
-            ['Advanced', 'Avanzado'],
-            ['Done', 'Hecho'],
-            ['Completed', 'Completado'],
-            ['Success', 'Éxito'],
-            ['Error', 'Error'],
-            ['Notice', 'Aviso'],
-            ['Update Available', 'Actualización disponible'],
-            ['Up to Date', 'Actualizado'],
-            ['Changelog', 'Registro de cambios'],
-            ['Download', 'Descargar'],
-            ['Later', 'Más tarde'],
-            ['Loading', 'Cargando'],
-            ['Month', 'Mes'],
-            ['Backup', 'Copia de seguridad'],
-            ['License', 'Licencia'],
-            ['Website', 'Sitio web'],
-            ['Optional', 'Opcional'],
-            ['Test', 'Probar'],
-            ['Last Sync', 'Última sincronización'],
-            ['Never', 'Nunca'],
-            ['Conflicts', 'Conflictos'],
-            ['No free time', 'Sin tiempo libre'],
-            ['Time conflict', 'Conflicto de tiempo'],
-            ['days', 'días'],
-        ],
-        hi: [
-            ['Settings', 'सेटिंग्स'],
-            ['General', 'सामान्य'],
-            ['Appearance', 'दिखावट'],
-            ['Language', 'भाषा'],
-            ['Projects', 'परियोजनाएँ'],
-            ['Project', 'परियोजना'],
-            ['Tasks', 'कार्य'],
-            ['Task', 'कार्य'],
-            ['Inbox', 'इनबॉक्स'],
-            ['Next Actions', 'अगले कदम'],
-            ['Next', 'अगला'],
-            ['Waiting', 'प्रतीक्षा'],
-            ['Someday/Maybe', 'कभी/शायद'],
-            ['Calendar', 'कैलेंडर'],
-            ['Review', 'समीक्षा'],
-            ['Archive', 'आर्काइव'],
-            ['Archived', 'आर्काइव्ड'],
-            ['Trash', 'ट्रैश'],
-            ['Add', 'जोड़ें'],
-            ['Delete', 'हटाएँ'],
-            ['Save', 'सहेजें'],
-            ['Cancel', 'रद्द करें'],
-            ['Search', 'खोजें'],
-            ['Notifications', 'सूचनाएँ'],
-            ['Sync', 'सिंक'],
-            ['Data', 'डेटा'],
-            ['About', 'परिचय'],
-            ['Advanced', 'उन्नत'],
-            ['Done', 'पूर्ण'],
-            ['Completed', 'पूर्ण'],
-            ['Success', 'सफल'],
-            ['Error', 'त्रुटि'],
-            ['Notice', 'सूचना'],
-            ['Update Available', 'अपडेट उपलब्ध'],
-            ['Up to Date', 'नवीनतम'],
-            ['Changelog', 'परिवर्तन लॉग'],
-            ['Download', 'डाउनलोड'],
-            ['Later', 'बाद में'],
-            ['Loading', 'लोड हो रहा है'],
-            ['Month', 'माह'],
-            ['Backup', 'बैकअप'],
-            ['License', 'लाइसेंस'],
-            ['Website', 'वेबसाइट'],
-            ['Optional', 'वैकल्पिक'],
-            ['Test', 'परीक्षण'],
-            ['Last Sync', 'पिछला सिंक'],
-            ['Never', 'कभी नहीं'],
-            ['Conflicts', 'संघर्ष'],
-            ['No free time', 'खाली समय नहीं'],
-            ['Time conflict', 'समय टकराव'],
-            ['days', 'दिन'],
-        ],
-        ar: [
-            ['Settings', 'الإعدادات'],
-            ['General', 'عام'],
-            ['Appearance', 'المظهر'],
-            ['Language', 'اللغة'],
-            ['Projects', 'المشاريع'],
-            ['Project', 'المشروع'],
-            ['Tasks', 'المهام'],
-            ['Task', 'مهمة'],
-            ['Inbox', 'الوارد'],
-            ['Next Actions', 'الخطوات التالية'],
-            ['Next', 'التالي'],
-            ['Waiting', 'انتظار'],
-            ['Someday/Maybe', 'يومًا ما/ربما'],
-            ['Calendar', 'التقويم'],
-            ['Review', 'المراجعة'],
-            ['Archive', 'الأرشيف'],
-            ['Archived', 'مؤرشف'],
-            ['Trash', 'المهملات'],
-            ['Add', 'إضافة'],
-            ['Delete', 'حذف'],
-            ['Save', 'حفظ'],
-            ['Cancel', 'إلغاء'],
-            ['Search', 'بحث'],
-            ['Notifications', 'الإشعارات'],
-            ['Sync', 'مزامنة'],
-            ['Data', 'البيانات'],
-            ['About', 'حول'],
-            ['Advanced', 'متقدم'],
-            ['Done', 'تم'],
-            ['Completed', 'مكتمل'],
-            ['Success', 'نجاح'],
-            ['Error', 'خطأ'],
-            ['Notice', 'تنبيه'],
-            ['Update Available', 'تحديث متاح'],
-            ['Up to Date', 'محدث'],
-            ['Changelog', 'سجل التغييرات'],
-            ['Download', 'تنزيل'],
-            ['Later', 'لاحقًا'],
-            ['Loading', 'جارٍ التحميل'],
-            ['Month', 'شهر'],
-            ['Backup', 'نسخة احتياطية'],
-            ['License', 'الترخيص'],
-            ['Website', 'الموقع'],
-            ['Optional', 'اختياري'],
-            ['Test', 'اختبار'],
-            ['Last Sync', 'آخر مزامنة'],
-            ['Never', 'أبدًا'],
-            ['Conflicts', 'تعارضات'],
-            ['No free time', 'لا يوجد وقت فراغ'],
-            ['Time conflict', 'تعارض في الوقت'],
-            ['days', 'أيام'],
-        ],
-        de: dePairs,
-        ru: ruPairs,
-        ja: jaPairs,
-        fr: frPairs,
-        pt: ptPairs,
-        pl: plPairs,
-        ko: koPairs,
-        it: itPairs,
-        tr: trPairs,
-        nl: nlPairs,
-        cs: csPairs,
-        vi: viPairs,
-};
-
-const escapeRegExp = (value: string) => value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-
-const buildReplacements = (pairs: Array<[string, string]>) =>
-        pairs
-            .filter(([from]) => from)
-            .sort((a, b) => b[0].length - a[0].length)
-            .map(([from, to]) => [new RegExp(escapeRegExp(from), 'gi'), to] as const);
-
-const PROTECTED_TOKENS = [
-        '/start:<when>',
-        '/due:<when>',
-        '/review:<when>',
-        '/note:<text>',
-        '/link:<url>',
-        '/next',
-        '/area:<name>',
-        '!Area',
-        '@context',
-        '#tag',
-        '+Project',
-        'project:',
-];
-
-export const autoTranslate = (value: string, lang: Language): string => {
-        const pairs = translationPairs[lang];
-        if (!pairs || pairs.length === 0) return value;
-
-        const placeholders = new Map<string, string>();
-        let protectedValue = value;
-        PROTECTED_TOKENS.forEach((token, index) => {
-                if (!protectedValue.includes(token)) return;
-                const key = `__MW_TOKEN_${index}__`;
-                placeholders.set(key, token);
-                protectedValue = protectedValue.split(token).join(key);
-        });
-
-        const replacements = buildReplacements(pairs);
-        let result = protectedValue;
-        for (const [pattern, replacement] of replacements) {
-                result = result.replace(pattern, replacement);
-        }
-        for (const [key, token] of placeholders.entries()) {
-                result = result.split(key).join(token);
-        }
-        return result;
-};
-
-export const translateText = (value: string, lang: Language): string => {
-        if (lang === 'en' || lang === 'zh' || lang === 'zh-Hant') return value;
-        return autoTranslate(value, lang);
-};
