--- packages/core/src/i18n/locales/i18n-types.ts
@@ -1,1 +1,0 @@
-export type Language = 'en' | 'vi' | 'zh' | 'zh-Hant' | 'es' | 'hi' | 'ar' | 'de' | 'ru' | 'ja' | 'fr' | 'pt' | 'pl' | 'ko' | 'cs' | 'it' | 'tr' | 'nl';
