+++ packages/core/src/i18n/locales/vi.ts
@@ -1,0 +1,1838 @@
+export const viOverrides: Record<string, string> = {
+        // App
+        'app.name': 'Mindwtr',
+        'markdown.copyCode': 'Sao chép mã',
+
+        // Navigation
+        'nav.main': 'Chính',
+        'nav.inbox': 'Hộp thư đến',
+        'nav.board': 'Bảng Kanban',
+        'nav.projects': 'Dự án',
+        'nav.contexts': 'Ngữ cảnh',
+        'nav.next': 'Hành động tiếp theo',
+        'nav.waiting': 'Đang chờ',
+        'nav.someday': 'Tương lai/Có thể',
+        'nav.reference': 'Tham khảo',
+        'nav.calendar': 'Lịch',
+        'nav.review': 'Xem lại',
+        'nav.agenda': 'Tập trung',
+        'nav.settings': 'Cài đặt',
+        'nav.done': 'Hoàn thành',
+        'nav.archived': 'Lưu trữ',
+        'nav.trash': 'Thùng rác',
+        'nav.addTask': 'Thêm nhiệm vụ',
+        'nav.sectionFocus': 'Tập trung',
+        'nav.sectionLists': 'Danh sách',
+        'nav.sectionOrganize': 'Sắp xếp',
+        'nav.sectionArchive': 'Lưu trữ',
+
+        // Quick Add
+        'quickAdd.help': 'Thêm nhanh hỗ trợ /start:<when>, /due:<when>, /review:<when>, /note:<text>, /link:<url>, /next, /area:<name> hoặc !Area, @context, #tag, +Project. Ngày đến hạn chỉ là ngày trừ khi bạn thêm giờ.',
+        'quickAdd.example': 'Ví dụ: Gọi cho mẹ /start:tomorrow /due:tomorrow 5pm @phone #family',
+        'quickAdd.placeholder': 'Thêm nhiệm vụ… dùng @ngữ cảnh +Dự án #thẻ',
+        'quickAdd.addAnother': 'Thêm cái khác',
+        'quickAdd.inputLabel': 'Tiêu đề nhiệm vụ',
+        'quickAdd.inputHint': 'Nhập tiêu đề nhiệm vụ, sau đó nhấn thêm hoặc phím Enter.',
+        'quickAdd.audioCaptureLabel': 'Ghi âm nhanh',
+        'quickAdd.audioRecord': 'Bắt đầu ghi âm',
+        'quickAdd.audioStop': 'Dừng ghi âm',
+        'quickAdd.audioRecording': 'Đang ghi âm...',
+        'quickAdd.audioNoteTitle': 'Ghi chú âm thanh',
+        'quickAdd.audioPermissionTitle': 'Cần quyền truy cập micrô',
+        'quickAdd.audioPermissionBody': 'Bật quyền truy cập micrô để ghi âm.',
+        'quickAdd.audioErrorTitle': 'Ghi âm thất bại',
+        'quickAdd.audioErrorBody': 'Không thể ghi âm. Vui lòng thử lại.',
+        'quickAdd.invalidDateCommand': 'Lệnh ngày không hợp lệ',
+        'quickAdd.bulkConfirmTitle': 'Tạo {{count}} nhiệm vụ?',
+        'quickAdd.bulkConfirmBody': 'Dòng trống sẽ bị bỏ qua. Mỗi dòng dùng cú pháp Thêm nhanh.',
+        'quickAdd.bulkConfirmCreate': 'Tạo nhiệm vụ',
+        'quickAdd.bulkMoreLines': '+{{count}} dòng nữa',
+        'quickAdd.bulkImportTextFile': 'Nhập .txt',
+        'quickAdd.bulkImportTextFileLabel': 'Nhập tệp văn bản',
+        'quickAdd.bulkImportError': 'Không thể đọc tệp văn bản đó.',
+        'quickAdd.bulkCreateError': 'Không thể tạo tất cả nhiệm vụ.',
+
+        // Checklist
+        'checklist.progress': 'Tiến độ danh sách kiểm tra',
+
+        // Keybindings (Desktop)
+        'keybindings.helpTitle': 'Phím tắt',
+        'keybindings.helpSubtitle': 'Nhấn Esc để đóng',
+        'keybindings.styleLabel': 'Kiểu phím tắt',
+        'keybindings.style.vim': 'Vim',
+        'keybindings.style.emacs': 'Emacs',
+        'keybindings.section.global': 'Toàn cục',
+        'keybindings.section.taskList': 'Danh sách nhiệm vụ',
+        'keybindings.openHelp': 'Hiện phím tắt',
+        'keybindings.openSettings': 'Mở cài đặt',
+        'keybindings.openSearch': 'Mở tìm kiếm',
+        'keybindings.quickAdd': 'Thêm nhanh',
+        'keybindings.toggleSidebar': 'Bật/tắt thanh bên',
+        'keybindings.toggleFocusMode': 'Bật/tắt chế độ tập trung',
+        'keybindings.toggleFullscreen': 'Bật/tắt toàn màn hình (ứng dụng máy tính)',
+        'keybindings.focusSidebar': 'Tập trung thanh bên',
+        'keybindings.focusContent': 'Tập trung nội dung chính',
+        'keybindings.goInbox': 'Đến Hộp thư đến',
+        'keybindings.goNext': 'Đến Hành động tiếp theo',
+        'keybindings.goAgenda': 'Đến Tập trung',
+        'keybindings.goProjects': 'Đến Dự án',
+        'keybindings.goContexts': 'Đến Ngữ cảnh',
+        'keybindings.goReview': 'Đến Xem lại',
+        'keybindings.goWaiting': 'Đến Đang chờ',
+        'keybindings.goSomeday': 'Đến Tương lai/Có thể',
+        'keybindings.goReference': 'Đến Tham khảo',
+        'keybindings.goCalendar': 'Đến Lịch',
+        'keybindings.goBoard': 'Đến Bảng Kanban',
+        'keybindings.goDone': 'Đến Hoàn thành',
+        'keybindings.goArchived': 'Đến Lưu trữ',
+        'keybindings.switchArea': 'Chuyển sang Khu vực 1-9',
+        'keybindings.clearAreaFilter': 'Xóa bộ lọc khu vực',
+        'keybindings.list.nextPrev': 'Di chuyển lựa chọn xuống/lên',
+        'keybindings.list.firstLast': 'Nhảy đến đầu/cuối',
+        'keybindings.list.edit': 'Chỉnh sửa nhiệm vụ đã chọn',
+        'keybindings.list.cancelEdit': 'Hủy chỉnh sửa',
+        'keybindings.list.saveEdit': 'Lưu chỉnh sửa',
+        'keybindings.list.toggleDone': 'Bật/tắt hoàn thành',
+        'keybindings.list.delete': 'Xóa nhiệm vụ đã chọn',
+        'keybindings.list.newTask': 'Tập trung ô nhập nhiệm vụ',
+        'keybindings.list.toggleDetails': 'Bật/tắt chi tiết danh sách',
+        'keybindings.list.toggleDensity': 'Bật/tắt mật độ',
+
+        // Tabs (Mobile)
+        'tab.inbox': 'Hộp thư đến',
+        'tab.next': 'Tập trung',
+        'tab.board': 'Bảng',
+        'tab.menu': 'Trình đơn',
+        'tab.agenda': 'Lịch trình',
+        'tab.calendar': 'Lịch',
+        'tab.review': 'Xem lại',
+
+        // List Views
+        'list.inbox': 'Hộp thư đến',
+        'list.todo': 'Cần làm',
+        'list.inProgress': 'Đang thực hiện',
+        'list.next': 'Hành động tiếp theo',
+        'list.someday': 'Tương lai/Có thể',
+        'list.reference': 'Tham khảo',
+        'list.waiting': 'Đang chờ',
+        'list.done': 'Đã hoàn thành',
+        'list.tasks': 'Danh sách nhiệm vụ',
+        'list.confirmBatchDelete': 'Xóa các nhiệm vụ đã chọn?',
+        'list.taskDeleted': 'Đã xóa nhiệm vụ',
+        'task.markedDone': '{title} đã được đánh dấu hoàn thành',
+        'list.tasksDeleted': 'Đã xóa {{count}} nhiệm vụ',
+        'list.noTasks': 'Không tìm thấy nhiệm vụ nào',
+        'list.filtering': 'Đang lọc...',
+        'list.details': 'Chi tiết',
+        'list.detailsOff': 'Tắt chi tiết',
+        'list.compact': 'Thu gọn',
+        'list.density': 'Mật độ',
+        'list.densityComfortable': 'Thoải mái',
+        'list.densityCompact': 'Thu gọn',
+        'list.groupBy': 'Nhóm theo',
+        'list.groupByNone': 'Không nhóm',
+        'list.groupByContext': 'Ngữ cảnh',
+        'list.groupByArea': 'Khu vực',
+        'reference.empty': 'Chưa có gì được lưu',
+        'reference.emptyHint': 'Tham khảo lưu thông tin bạn có thể cần sau — không cần hành động.',
+
+        // Status Labels
+        'status.inbox': 'Hộp thư đến',
+        'status.todo': 'Cần làm',
+        'status.next': 'Tiếp theo',
+        'status.in-progress': 'Đang thực hiện',
+        'status.waiting': 'Đang chờ',
+        'status.someday': 'Tương lai',
+        'status.reference': 'Tham khảo',
+        'status.done': 'Hoàn thành',
+        'status.archived': 'Đã lưu trữ',
+        'status.active': 'Đang hoạt động',
+
+        // Priority Labels
+        'priority.low': 'Thấp',
+        'priority.medium': 'Trung bình',
+        'priority.high': 'Cao',
+        'priority.urgent': 'Khẩn cấp',
+        'energyLevel.low': 'Năng lượng thấp',
+        'energyLevel.medium': 'Năng lượng trung bình',
+        'energyLevel.high': 'Năng lượng cao',
+
+        // Task Edit
+        'taskEdit.editTask': 'Chỉnh sửa nhiệm vụ',
+        'taskEdit.checklist': 'Danh sách kiểm tra',
+        'taskEdit.tab.task': 'Nhiệm vụ',
+        'taskEdit.tab.list': 'Danh sách',
+        'taskEdit.tab.view': 'Xem',
+        'taskEdit.title': 'Nhiệm vụ',
+        'taskEdit.editorLayoutHelpLabel': 'Trợ giúp bố cục trình chỉnh sửa',
+        'taskEdit.editorLayoutHelpText': 'Bạn có thể tùy chỉnh các trường hiển thị ở đây trong Cài đặt -> GTD -> Bố cục trình chỉnh sửa nhiệm vụ.',
+        'taskEdit.start': 'Bắt đầu',
+        'taskEdit.shoppingListPrompt': 'Danh sách mua sắm?',
+        'taskEdit.openChecklistMode': 'Mở chế độ danh sách kiểm tra',
+        'taskEdit.addItem': 'Thêm mục',
+        'taskEdit.resetChecklist': 'Đặt lại danh sách kiểm tra',
+        'task.convertToReference': 'Chuyển thành tham khảo',
+        'taskEdit.duplicateTask': 'Nhân đôi nhiệm vụ',
+        'taskEdit.duplicateDoneTitle': 'Đã nhân đôi nhiệm vụ',
+        'taskEdit.duplicateDoneBody': 'Một bản sao mới đã được tạo trong Hộp thư đến.',
+        'taskEdit.aiClarify': 'AI làm rõ',
+        'taskEdit.aiBreakdown': 'AI phân tích',
+        'taskEdit.itemNamePlaceholder': 'Tên mục',
+        'taskEdit.titleLabel': 'Tiêu đề',
+        'taskEdit.statusLabel': 'Trạng thái',
+        'taskEdit.priorityLabel': 'Ưu tiên',
+        'taskEdit.energyLevel': 'Mức năng lượng',
+        'taskEdit.assignedTo': 'Giao cho',
+        'taskEdit.assignedToPlaceholder': 'Đang chờ ai?',
+        'taskEdit.contextsLabel': 'Ngữ cảnh',
+        'taskEdit.contextsPlaceholder': '@nhà, @công việc',
+        'taskEdit.tagsLabel': 'Thẻ',
+        'taskEdit.tagsPlaceholder': '#khẩn cấp, #ý tưởng',
+        'taskEdit.timeEstimateLabel': 'Ước tính thời gian',
+        'taskEdit.recurrenceLabel': 'Lặp lại',
+        'taskEdit.descriptionLabel': 'Mô tả',
+        'taskEdit.descriptionPlaceholder': 'Thêm ghi chú...',
+        'taskEdit.textDirectionLabel': 'Hướng văn bản',
+        'taskEdit.textDirection.auto': 'Tự động',
+        'taskEdit.textDirection.ltr': 'Trái sang phải',
+        'taskEdit.textDirection.rtl': 'Phải sang trái',
+        'taskEdit.basic': 'Cơ bản',
+        'taskEdit.scheduling': 'Lên lịch',
+        'taskEdit.schedulingEmpty': 'Tất cả các trường lên lịch đều bị ẩn. Hiện chúng trong Cài đặt > GTD > Bố cục trình chỉnh sửa nhiệm vụ.',
+        'taskEdit.organization': 'Tổ chức',
+        'taskEdit.organizationEmpty': 'Tất cả các trường tổ chức đều bị ẩn. Hiện chúng trong Cài đặt > GTD > Bố cục trình chỉnh sửa nhiệm vụ.',
+        'taskEdit.details': 'Chi tiết',
+        'taskEdit.detailsEmpty': 'Tất cả các trường chi tiết đều bị ẩn. Hiện chúng trong Cài đặt > GTD > Bố cục trình chỉnh sửa nhiệm vụ.',
+        'markdown.preview': 'Xem trước',
+        'markdown.edit': 'Chỉnh sửa',
+        'markdown.expand': 'Mở rộng',
+        'markdown.collapse': 'Thu gọn',
+        'markdown.referenceDeletedTask': 'nhiệm vụ đã xóa',
+        'markdown.referenceDeletedProject': 'dự án đã xóa',
+        'markdown.referenceRestore': 'Khôi phục',
+        'markdown.toolbar.heading': 'Chèn tiêu đề',
+        'markdown.toolbar.bold': 'In đậm',
+        'markdown.toolbar.italic': 'In nghiêng',
+        'markdown.toolbar.strikethrough': 'Gạch ngang',
+        'markdown.toolbar.bulletList': 'Danh sách dấu đầu dòng',
+        'markdown.toolbar.horizontalRule': 'Đường kẻ ngang',
+        'markdown.toolbar.link': 'Chèn liên kết',
+        'markdown.toolbar.code': 'Mã nội tuyến',
+        'markdown.toolbar.codeBlock': 'Khối mã',
+        'attachments.title': 'Tệp đính kèm',
+        'attachments.addFile': 'Thêm tệp',
+        'attachments.addPhoto': 'Thêm ảnh',
+        'attachments.addLink': 'Thêm liên kết',
+        'attachments.open': 'Mở',
+        'attachments.download': 'Tải xuống',
+        'attachments.missing': 'Tệp bị thiếu',
+        'attachments.remove': 'Xóa',
+        'attachments.transferProgress': 'Tiến độ truyền tệp đính kèm',
+        'attachments.linkPlaceholder': 'https://example.com',
+        'attachments.linkInputHint': 'Mẹo: dán URL, hoặc dùng "Tiêu đề | URL".',
+        'attachments.fileNotSupported': 'Tệp đính kèm chỉ được hỗ trợ trên ứng dụng máy tính.',
+        'attachments.fileTooLarge': 'Tệp quá lớn để tải lên.',
+        'attachments.invalidFileType': 'Loại tệp không được hỗ trợ.',
+        'attachments.invalidLink': 'Vui lòng nhập URL hợp lệ.',
+        'attachments.photoUnavailableTitle': 'Bộ chọn ảnh không khả dụng',
+        'attachments.photoUnavailableBody': 'Xây dựng lại ứng dụng để bật tệp đính kèm ảnh.',
+        'taskEdit.locationLabel': 'Địa điểm',
+        'taskEdit.locationPlaceholder': 'vd. Văn phòng',
+        'taskEdit.projectLabel': 'Dự án',
+        'taskEdit.noProjectOption': 'Không có dự án',
+        'taskEdit.sectionLabel': 'Phần dự án',
+        'taskEdit.noSectionOption': 'Không có phần',
+        'taskEdit.areaLabel': 'Khu vực',
+        'taskEdit.noAreaOption': 'Không có khu vực',
+        'taskEdit.moreOptions': 'Thêm tùy chọn',
+        'taskEdit.hideOptions': 'Ẩn tùy chọn',
+        'taskEdit.discardChanges': 'Bỏ các thay đổi chưa lưu?',
+        'taskEdit.discardChangesDesc': 'Các thay đổi của bạn sẽ bị mất nếu bạn rời đi ngay bây giờ.',
+        'taskEdit.startDateLabel': 'Ngày bắt đầu',
+        'taskEdit.dueDateLabel': 'Ngày đến hạn',
+        'taskEdit.reviewDateLabel': 'Ngày xem lại',
+        'taskEdit.repeatReminderLabel': 'Nhắc nhở lặp lại',
+        'taskEdit.repeatReminderOff': 'Tắt',
+        'taskEdit.repeatReminderEveryMinutes': 'Mỗi {count} phút',
+        'taskEdit.repeatReminderMinutesShort': '{count} phút',
+        'taskEdit.repeatReminderHint': 'Thông báo lại theo khoảng thời gian này cho đến khi bạn hoàn thành hoặc hoãn, tối đa một giới hạn.',
+        'taskEdit.pushCountHint': 'Đã đẩy',
+        'taskEdit.unblocksLabel': 'Mở khóa',
+        'taskStatus.changeStatus': 'Thay đổi trạng thái',
+        'task.toggleDetails': 'Bật/tắt chi tiết nhiệm vụ',
+        'task.hoverHint': 'Nhấp để bật/tắt chi tiết / Nhấp đúp để chỉnh sửa',
+        'task.select': 'Chọn nhiệm vụ',
+        'task.aria.delete': 'Xóa nhiệm vụ',
+        'task.deleteConfirmBody': 'Chuyển nhiệm vụ này vào thùng rác?',
+        'task.aria.status': 'Trạng thái nhiệm vụ',
+        'task.aria.description': 'Mô tả nhiệm vụ',
+        'task.aria.startDate': 'Ngày bắt đầu',
+        'task.aria.startTime': 'Giờ bắt đầu',
+        'task.aria.reviewDate': 'Ngày xem lại',
+        'task.aria.reviewTime': 'Giờ xem lại',
+        'task.aria.recurrence': 'Lặp lại',
+        'task.aria.timeEstimate': 'Ước tính thời gian',
+        'task.aria.contexts': 'Ngữ cảnh',
+        'task.aria.tags': 'Thẻ',
+        'task.aria.area': 'Khu vực',
+        'task.aria.dueDate': 'Ngày đến hạn',
+        'task.aria.dueTime': 'Giờ đến hạn',
+        'task.aria.location': 'Địa điểm',
+
+        // Recurrence Labels
+        'recurrence.none': 'Không',
+        'recurrence.daily': 'Hàng ngày',
+        'recurrence.weekly': 'Hàng tuần',
+        'recurrence.monthly': 'Hàng tháng',
+        'recurrence.yearly': 'Hàng năm',
+        'recurrence.custom': 'Tùy chỉnh…',
+        'recurrence.customTitle': 'Lặp lại tùy chỉnh',
+        'recurrence.repeatEvery': 'Lặp lại mỗi',
+        'recurrence.repeatOn': 'Lặp lại vào',
+        'recurrence.endsLabel': 'Kết thúc',
+        'recurrence.endsNever': 'Không bao giờ',
+        'recurrence.endsOnDate': 'Vào ngày',
+        'recurrence.endsAfterCount': 'Sau',
+        'recurrence.occurrenceUnit': 'lần',
+        'recurrence.dayUnit': 'ngày',
+        'recurrence.weekUnit': 'tuần',
+        'recurrence.monthUnit': 'tháng',
+        'recurrence.onLabel': 'Vào',
+        'recurrence.onDayOfMonth': 'Ngày {day}',
+        'recurrence.onNthWeekday': '{ordinal} {weekday}',
+        'recurrence.ordinalSelectLabel': 'Thứ tự lặp lại',
+        'recurrence.weekdaySelectLabel': 'Ngày trong tuần lặp lại',
+        'recurrence.ordinal.first': 'Thứ nhất',
+        'recurrence.ordinal.second': 'Thứ hai',
+        'recurrence.ordinal.third': 'Thứ ba',
+        'recurrence.ordinal.fourth': 'Thứ tư',
+        'recurrence.ordinal.last': 'Cuối cùng',
+        'recurrence.monthlyOnDay': 'Cùng ngày mỗi tháng',
+        'recurrence.monthlyOnLastWeekday': '{weekday} cuối cùng',
+        'recurrence.strategyLabel': 'Chiến lược',
+        'recurrence.strategyStrict': 'Chặt chẽ',
+        'recurrence.strategyFluid': 'Linh hoạt',
+        'recurrence.strategyStrictDesc': 'Đến hạn vào ngày đã lên lịch (vd. hóa đơn)',
+        'recurrence.strategyFluidDesc': 'Đến hạn sau khi hoàn thành (vd. giặt quần áo)',
+        'recurrence.afterCompletion': 'Lặp lại sau khi hoàn thành',
+        'recurrence.afterCompletionShort': 'sau khi hoàn thành',
+        'recurrence.showFutureInCalendar': 'Hiện lần xuất hiện tiếp theo trong Lịch',
+        'recurrence.showFutureInCalendarHint': 'Chỉ xem trước cho lên kế hoạch; nhiệm vụ tiếp theo vẫn được tạo khi hoàn thành cái này.',
+
+        // Inbox (Mobile)
+        'inbox.title': 'Hộp thư đến',
+        'inbox.processButton': 'Xử lý hộp thư đến',
+        'inbox.addPlaceholder': 'Thêm nhiệm vụ vào hộp thư đến...',
+        'inbox.empty': 'Không có nhiệm vụ trong hộp thư đến',
+        'inbox.emptyAddHint': 'Ghi lại mọi thứ trong đầu bạn ở đây — sắp xếp sau.',
+        'inbox.emptyAddHintVoice': 'Hộp thư đến trống. Nhấn micrô để ghi lại.',
+        'mobileHelp.title': 'Mẹo di động',
+        'mobileHelp.gesturesTitle': 'Cử chỉ',
+        'mobileHelp.swipeRight': 'Vuốt phải trên nhiệm vụ để chuyển tiếp.',
+        'mobileHelp.swipeLeft': 'Vuốt trái để xóa, khôi phục hoặc gửi lại.',
+        'mobileHelp.longPressCapture': 'Nhấn giữ + để dùng chế độ ghi lại thay thế.',
+        'mobileHelp.shortcutsTitle': 'Phím tắt',
+        'mobileHelp.appShortcuts': 'Nhấn giữ biểu tượng ứng dụng để dùng phím tắt Thêm nhiệm vụ, Tập trung và Lịch.',
+        'inbox.refineTitle': 'Tinh chỉnh nhiệm vụ này',
+        'inbox.refineHint': 'Làm rõ tiêu đề và chi tiết trước khi quyết định làm gì tiếp theo.',
+        'inbox.refineNext': 'Tiếp theo',
+        'inbox.refineDelete': 'Xóa',
+        'inbox.isActionable': 'Đây có thể thực hiện không?',
+        'inbox.actionableHint': 'Bạn có thể thực hiện hành động vật lý trên cái này không?',
+        'inbox.yes': 'Có',
+        'inbox.no': 'Không',
+        'inbox.trash': 'Thùng rác',
+        'inbox.someday': 'Tương lai',
+        'inbox.twoMinRule': 'Liệu có mất dưới 2 phút không?',
+        'inbox.twoMinHint': 'Nếu có, hãy làm ngay!',
+        'inbox.doneIt': 'Xong rồi!',
+        'inbox.takesLonger': 'Mất hơn 2 phút',
+        'inbox.whatNext': 'Tiếp theo là gì?',
+        'inbox.whoShouldDoIt': 'Ai nên làm điều này?',
+        'inbox.illDoIt': 'Tôi sẽ làm',
+        'inbox.delegate': 'Giao cho người khác',
+        'inbox.whereDoIt': 'Bạn sẽ làm điều này ở đâu?',
+        'inbox.addContext': 'Thêm ngữ cảnh để tìm lại sau',
+        'inbox.yesActionable': 'Có, có thể thực hiện',
+        'inbox.selectMultipleHint': '(Chọn nhiều hoặc không chọn)',
+        'inbox.selectedLabel': 'Đã chọn:',
+        'inbox.addContextPlaceholder': 'Thêm ngữ cảnh mới...',
+        'inbox.waitingQuestion': 'Bạn đang chờ ai/điều gì?',
+        'inbox.waitingHint': 'Thêm ghi chú để nhớ bạn đang chờ gì',
+        'inbox.waitingPlaceholder': 'vd. Đang chờ John xem xét...',
+        'inbox.assignProjectQuestion': 'Gán vào dự án? (Tùy chọn)',
+        'inbox.noProject': 'Không có dự án',
+        'inbox.skip': 'Bỏ qua',
+
+        // Next Actions (Mobile)
+        'next.title': 'Hành động tiếp theo',
+        'next.ready': 'nhiệm vụ sẵn sàng',
+        'next.current': 'Hành động hiện tại',
+        'next.promote': 'Sẵn sàng thăng cấp',
+        'next.promoteHint': 'Nhấn mũi tên để chuyển sang Hành động tiếp theo',
+        'next.noTasks': 'Không có nhiệm vụ cần làm. Thêm vào Hộp thư đến và xử lý trước.',
+        'next.noContext': 'Không có Hành động tiếp theo với',
+        'next.warningCount': 'mục trong Hành động tiếp theo',
+        'next.warningHint': 'Hãy xem xét tập trung vào ít dự án hơn. GTD khuyến nghị giữ Hành động tiếp theo trong khoảng 10-15 mục để rõ ràng hơn.',
+
+        // Contact / Contexts (Mobile updates)
+        'contexts.title': 'Ngữ cảnh',
+        'contexts.filter': 'Lọc nhiệm vụ theo ngữ cảnh',
+        'filters.label': 'Bộ lọc',
+        'filters.clear': 'Xóa',
+        'filters.show': 'Hiện',
+        'filters.hide': 'Ẩn',
+        'filters.hideFutureTasks': 'Ẩn nhiệm vụ tương lai',
+        'filters.contexts': 'Ngữ cảnh & thẻ',
+        'filters.priority': 'Ưu tiên',
+        'filters.timeEstimate': 'Ước tính thời gian',
+        'filters.after': 'Sau',
+        'filters.before': 'Trước',
+        'filters.hasDescription': 'Có mô tả',
+        'filters.noDescription': 'Không có mô tả',
+        'filters.starred': 'Đã đánh dấu sao',
+        'filters.notStarred': 'Chưa đánh dấu sao',
+        'filters.datePreset.today': 'Hôm nay',
+        'filters.datePreset.this_week': 'Tuần này',
+        'filters.datePreset.this_month': 'Tháng này',
+        'filters.datePreset.overdue': 'Quá hạn',
+        'filters.datePreset.no_date': 'Không có ngày',
+        'quickDate.today': 'Hôm nay',
+        'quickDate.tomorrow': 'Ngày mai',
+        'quickDate.in3Days': '+3 ngày',
+        'quickDate.nextWeek': 'Tuần tới',
+        'quickDate.nextMonth': 'Tháng tới',
+        'quickDate.noDate': 'Không có ngày',
+        'filters.projects': 'Dự án',
+        'filters.noMatch': 'Không có nhiệm vụ nào khớp với các bộ lọc này.',
+        'contexts.all': 'Tất cả ngữ cảnh',
+        'contexts.none': 'Không có ngữ cảnh',
+        'contexts.noContexts': 'Không tìm thấy ngữ cảnh nào. Thêm ngữ cảnh như @nhà, @côngviệc, @máytính vào nhiệm vụ của bạn',
+        'contexts.noTasks': 'Không có nhiệm vụ đang hoạt động cho ngữ cảnh này',
+
+        // Board
+        'board.title': 'Bảng Kanban',
+        'board.next': 'Hành động tiếp theo',
+        'board.todo': 'Cần làm',
+        'board.inProgress': 'Đang thực hiện',
+        'board.done': 'Hoàn thành',
+        'board.noTasks': 'Không có nhiệm vụ',
+        'board.hint': 'Giữ để kéo • Vuốt trái để xóa',
+        'board.dragTask': 'Kéo nhiệm vụ',
+        'board.delete': 'Xóa',
+
+        // Calendar
+        'calendar.title': 'Lịch',
+        'calendar.tasksAndEvents': 'Nhiệm vụ và sự kiện bên ngoài theo ngày',
+        'calendar.today': 'Hôm nay',
+        'calendar.day': 'Ngày',
+        'calendar.week': 'Tuần',
+        'calendar.month': 'Tháng',
+        'calendar.schedule': 'Lịch trình',
+        'calendar.prevMonth': 'Tháng trước',
+        'calendar.nextMonth': 'Tháng sau',
+        'calendar.year': 'Năm',
+        'calendar.searchMatches': '{count} kết quả trong chế độ xem này',
+        'calendar.noSearchMatches': 'Không có mục lịch nào khớp trong chế độ xem này',
+        'calendar.visibleCalendars': 'Lịch',
+        'calendar.openDayView': 'Mở chế độ xem ngày',
+        'calendar.more': 'thêm',
+        'calendar.time': 'Thời gian',
+        'calendar.items': 'mục',
+        'calendar.tasks': 'Nhiệm vụ',
+        'calendar.scheduleHelp': 'Tìm nhiệm vụ và đặt thời gian lịch.',
+        'calendar.planningTitle': 'Lên kế hoạch hành động tiếp theo',
+        'calendar.planningHelp': 'Chọn hành động tiếp theo hiện có và sắp xếp xung quanh lịch cố định của bạn.',
+        'calendar.planningForDate': 'Lên kế hoạch cho {date}',
+        'calendar.planningEmpty': 'Không có hành động tiếp theo chưa được lên lịch.',
+        'calendar.selectDayToPlan': 'Chọn ngày để lên kế hoạch trước.',
+        'calendar.addToCalendar': 'Thêm vào lịch',
+        'calendar.newTask': 'Nhiệm vụ mới',
+        'calendar.existingTask': 'Nhiệm vụ hiện có',
+        'calendar.taskTitle': 'Tiêu đề nhiệm vụ',
+        'calendar.findTask': 'Tìm nhiệm vụ',
+        'calendar.noMatchingTasks': 'Không có nhiệm vụ nào khớp',
+        'calendar.date': 'Ngày',
+        'calendar.start': 'Bắt đầu',
+        'calendar.end': 'Kết thúc',
+        'calendar.duration': 'Thời lượng',
+        'calendar.addTask': 'Thêm nhiệm vụ mới...',
+        'calendar.schedulePlaceholder': 'Tìm kiếm nhiệm vụ để lên lịch...',
+        'calendar.scheduleResults': 'Lên lịch',
+        'calendar.scheduleAction': 'Lên lịch',
+        'calendar.events': 'Sự kiện',
+        'calendar.allDay': 'Cả ngày',
+        'calendar.deadline': 'Hạn chót',
+        'calendar.projectedRecurrence': 'Dự kiến',
+        'calendar.projectedRecurrenceDescription': 'Chỉ xem trước cho lên kế hoạch. Hoàn thành nhiệm vụ lặp lại hiện tại để tạo nhiệm vụ thực sự tiếp theo.',
+        'calendar.unschedule': 'Xóa khỏi lịch',
+        'calendar.changeTime': 'Thay đổi giờ',
+        'calendar.noTasks': 'Không có nhiệm vụ cho ngày này',
+        'calendar.noFreeTime': 'Không có thời gian rảnh cho ngày này.',
+        'calendar.overlapWarning': 'Thời gian đó trùng với một sự kiện. Vui lòng chọn khung giờ rảnh.',
+        'calendar.invalidTimeRange': 'Chọn thời gian bắt đầu và kết thúc hợp lệ.',
+        'calendar.enterTaskTitle': 'Nhập tiêu đề nhiệm vụ.',
+        'calendar.taskTitleRequired': 'Nhập tiêu đề nhiệm vụ.',
+        'calendar.chooseTask': 'Chọn một nhiệm vụ.',
+        'calendar.taskRequired': 'Chọn một nhiệm vụ.',
+        'calendar.saveTaskFailed': 'Không thể lưu nhiệm vụ.',
+        'calendar.saveFailed': 'Không thể lưu nhiệm vụ.',
+        'calendar.createTaskFromEvent': 'Tạo nhiệm vụ',
+        'calendar.eventFallbackTitle': 'Sự kiện lịch',
+        'calendar.eventTaskCreatedTitle': 'Đã tạo nhiệm vụ',
+        'calendar.eventTaskCreated': 'Đã tạo nhiệm vụ từ sự kiện.',
+        'calendar.openInCalendar': 'Mở trong lịch',
+        'calendar.noFreeTimeTitle': 'Không có thời gian rảnh',
+        'calendar.timeConflictTitle': 'Xung đột thời gian',
+        'calendar.cannotOpenEventTitle': 'Không thể mở sự kiện',
+        'calendar.openDeviceEventOnly': 'Chỉ có thể mở sự kiện lịch thiết bị trong ứng dụng lịch.',
+        'calendar.openUnsupported': 'Ứng dụng lịch này không hỗ trợ mở sự kiện.',
+        'calendar.openFromCalendarApp': 'Mở sự kiện từ ứng dụng lịch của bạn.',
+        'share.unavailable': 'Chia sẻ không khả dụng',
+        'share.readFailed': 'Mindwtr không thể đọc văn bản hoặc URL từ mục được chia sẻ.',
+        'shortcuts.captureUnavailable': 'Phím tắt ghi lại không khả dụng',
+        'shortcuts.missingTitle': 'Mindwtr không thể đọc tiêu đề nhiệm vụ từ liên kết phím tắt đó.',
+
+        // Projects
+        'projects.title': 'Dự án',
+        'projects.statusLabel': 'Trạng thái',
+        'projects.activeSection': 'Dự án đang hoạt động',
+        'projects.noProjects': 'Chưa có dự án nào.',
+        'projects.noProjectsInArea': 'Không có dự án trong khu vực này.',
+        'projects.selectProject': 'Chọn dự án để xem nhiệm vụ',
+        'projects.addTask': 'Thêm nhiệm vụ',
+        'projects.addTaskPlaceholder': 'Thêm nhiệm vụ vào dự án này...',
+        'projects.noActiveTasks': 'Không có nhiệm vụ đang hoạt động trong dự án này.',
+        'projects.projectName': 'Tên dự án',
+        'projects.color': 'Màu sắc',
+        'projects.colorBlue': 'Xanh dương',
+        'projects.colorGreen': 'Xanh lá',
+        'projects.colorAmber': 'Hổ phách',
+        'projects.colorRed': 'Đỏ',
+        'projects.colorPurple': 'Tím',
+        'projects.colorPink': 'Hồng',
+        'projects.create': 'Tạo',
+        'areas.create': 'Tạo',
+        'areas.deleteConfirm': 'Xóa khu vực này? Các dự án và nhiệm vụ trong khu vực này sẽ được giữ lại và chuyển sang chưa phân công.',
+        'projects.deleteConfirm': 'Xóa dự án này? Các nhiệm vụ trong dự án này sẽ được giữ lại và chuyển sang chưa phân công.',
+        'projects.deleted': 'Dự án đã chuyển vào Thùng rác',
+        'projects.deleteFailed': 'Xóa dự án thất bại',
+        'projects.restoreFailed': 'Khôi phục dự án thất bại',
+        'projects.count': 'dự án',
+        'projects.add': 'Thêm',
+        'projects.addPlaceholder': 'Thêm dự án mới...',
+        'projects.emptyHint': 'Dự án nhóm các hành động tiếp theo liên quan hướng đến một kết quả. Tạo một dự án để bắt đầu.',
+        'projects.emptyHintFiltered': 'Thử chuyển bộ lọc Khu vực hoặc tạo dự án trong khu vực này.',
+        'projects.empty': 'Chưa có dự án nào',
+        'projects.back': '← Quay lại',
+        'project.notes': 'Ghi chú dự án',
+        'projects.notesPlaceholder': 'Thêm ngữ cảnh, kế hoạch hoặc ghi chú tham khảo cho dự án này...',
+        'projects.sectionNotes': 'Ghi chú phần',
+        'projects.sectionNotesPlaceholder': 'Thêm ghi chú cho phần này...',
+        'projects.reviewAt': 'Ngày xem lại',
+        'projects.areaLabel': 'Khu vực',
+        'projects.areaPlaceholder': 'vd. Công việc',
+        'projects.sectionsLabel': 'Các phần',
+        'projects.addSection': 'Thêm phần',
+        'projects.sectionPlaceholder': 'Tiêu đề phần',
+        'projects.noSection': 'Không có phần',
+        'projects.deleteSectionConfirm': 'Bạn có chắc muốn xóa phần này không?',
+        'projects.areaFilter': 'Bộ lọc khu vực',
+        'projects.allAreas': 'Tất cả khu vực',
+        'projects.noArea': 'Không có khu vực',
+        'projects.areaReorderFailed': 'Sắp xếp lại khu vực thất bại',
+        'projects.moveProjectFailed': 'Di chuyển dự án thất bại',
+        'projects.taskOrder': 'Thứ tự nhiệm vụ',
+        'projects.reorderTasks': 'Sắp xếp',
+        'projects.taskReorderFailed': 'Sắp xếp lại nhiệm vụ thất bại.',
+        'projects.deletedAreaFilterResetAlert': 'Khu vực đã chọn bị xóa, bộ lọc đã đặt lại về Tất cả khu vực.',
+        'projects.search': 'Tìm kiếm dự án',
+        'sections.search': 'Tìm kiếm phần',
+        'areas.search': 'Tìm kiếm khu vực',
+        'areas.manage': 'Khu vực',
+        'areas.new': 'Khu vực mới',
+        'areas.namePlaceholder': 'Tên khu vực',
+        'projects.sortAreas': 'Sắp xếp khu vực',
+        'projects.sortByName': 'Sắp xếp A–Z',
+        'projects.sortByColor': 'Sắp xếp theo màu',
+        'projects.moveUp': 'Di chuyển lên',
+        'projects.moveDown': 'Di chuyển xuống',
+        'projects.tagFilter': 'Bộ lọc thẻ',
+        'projects.allTags': 'Tất cả thẻ',
+        'projects.noTags': 'Không có thẻ',
+        'projects.deferredSection': 'Tương lai / Đang chờ',
+        'projects.duplicate': 'Nhân đôi',
+        'projects.complete': 'Hoàn thành',
+        'projects.archive': 'Lưu trữ',
+        'projects.reactivate': 'Kích hoạt lại',
+        'projects.completeConfirm': 'Đánh dấu dự án này hoàn thành và kết thúc tất cả các nhiệm vụ của nó?',
+        'projects.archiveConfirm': 'Lưu trữ dự án này và chuyển các nhiệm vụ chưa hoàn thành sang Hoàn thành?',
+        'projects.sequential': 'Tuần tự',
+        'projects.parallel': 'Song song',
+        'projects.sequentialTooltip': 'Tuần tự: một bước khả dụng được hiển thị; các bước sau chờ theo thứ tự.',
+        'projects.parallelTooltip': 'Song song: nhiều nhiệm vụ Tiếp theo độc lập có thể được hiển thị.',
+        'projects.availableNextAction': 'Hành động tiếp theo khả dụng',
+        'projects.laterInSequence': 'Sau trong chuỗi',
+        'projects.projectTypeHelpLabel': 'Trợ giúp loại dự án',
+        'projects.projectTypeHelpText': 'Dự án tuần tự hiển thị một hành động khả dụng tại một thời điểm. Dự án song song có thể hiển thị nhiều nhiệm vụ Tiếp theo độc lập.',
+        'projects.sequentialScope': 'Phạm vi tuần tự',
+        'projects.sequentialScopeHelpLabel': 'Trợ giúp phạm vi tuần tự',
+        'projects.sequentialScopeHelpText': 'Xuyên các phần hiển thị một hành động khả dụng cho toàn bộ dự án. Trong các phần hiển thị một hành động khả dụng cho mỗi phần.',
+        'projects.sequentialAcrossSections': 'Xuyên các phần',
+        'projects.sequentialWithinSections': 'Trong các phần',
+        'projects.reviewAtHint': 'Ngày nhắc nhở để xem lại dự án này.',
+        'projects.noNextAction': 'Không có hành động tiếp theo',
+        'projects.nextActionPromptTitle': 'Hành động tiếp theo là gì?',
+        'projects.nextActionPromptDesc': 'Chọn hoặc thêm hành động tiếp theo cho {{project}}.',
+        'projects.nextActionPromptChooseExisting': 'Chọn nhiệm vụ hiện có',
+        'projects.nextActionPromptAddNew': 'Thêm hành động tiếp theo mới',
+        'projects.nextActionPromptPlaceholder': 'Hành động tiếp theo mới...',
+        'projects.nextActionPromptAddButton': 'Thêm hành động tiếp theo',
+
+        // Review Steps
+        'review.title': 'Xem lại',
+        'review.intro': 'Làm rõ, cập nhật và sáng tạo.',
+        'review.openGuide': 'Xem lại hàng tuần',
+        'dailyReview.openGuide': 'Xem lại hàng ngày',
+        'dailyReview.title': 'Xem lại hàng ngày',
+        'dailyReview.introTitle': 'Khởi động hàng ngày',
+        'dailyReview.introDesc': 'Kiểm tra nhanh để luôn cập nhật.',
+        'dailyReview.start': 'Bắt đầu',
+        'dailyReview.todayStep': 'Hôm nay & Lịch',
+        'dailyReview.todayDesc': 'Kiểm tra những gì đến hạn hôm nay và các cam kết lịch của bạn.',
+        'dailyReview.focusStep': 'Tập trung hôm nay',
+        'dailyReview.focusDesc': 'Tùy chọn: đánh dấu các nhiệm vụ để hiển thị trong Tập trung hôm nay.',
+        'dailyReview.focusSelected': 'đã tập trung',
+        'dailyReview.inboxStep': 'Xử lý Hộp thư đến',
+        'dailyReview.inboxDesc': 'Làm rõ các đầu vào mới vào đúng danh sách.',
+        'dailyReview.waitingStep': 'Đang chờ',
+        'dailyReview.followUpToday': 'Theo dõi hôm nay',
+        'dailyReview.waitingDesc': 'Theo dõi bất cứ điều gì đến hạn.',
+        'dailyReview.completeTitle': 'Sẵn sàng',
+        'dailyReview.completeDesc': 'Bạn đã rõ ràng và sẵn sàng làm việc.',
+        'dailyReview.openAgenda': 'Mở Lịch trình',
+        'dailyReview.openInbox': 'Mở Hộp thư đến',
+        'dailyReview.openWaiting': 'Mở Đang chờ',
+        'dailyReview.openCalendar': 'Mở Lịch',
+        'review.timeFor': 'Đến lúc xem lại hàng tuần',
+        'review.timeForDesc': 'Giải phóng tâm trí và sắp xếp. Quy trình này sẽ hướng dẫn bạn dọn dẹp danh sách và lên kế hoạch cho tuần tới.',
+        'review.startReview': 'Bắt đầu xem lại',
+        'review.inboxStep': 'Xử lý Hộp thư đến',
+        'review.inboxStepDesc': 'Làm rõ và sắp xếp các mục trong hộp thư đến.',
+        'review.inboxZero': 'Mục tiêu Hộp thư đến bằng 0',
+        'review.inboxZeroDesc': 'mục trong Hộp thư đến của bạn. Xử lý chúng bằng cách làm rõ chúng là gì và sắp xếp vào hành động tiếp theo, dự án hoặc thùng rác.',
+        'review.inboxEmpty': 'Hộp thư đến trống! Làm tốt lắm.',
+        'review.inboxGuide': 'Xử lý từng mục: xóa, giao, đặt hành động tiếp theo hoặc chuyển sang Tương lai. Mục tiêu: hộp thư đến bằng 0!',
+        'review.calendarStep': 'Xem lại Lịch',
+        'review.calendarStepDesc': 'Xem lại lịch của bạn trong 7 ngày tới.',
+        'review.calendarEmpty': 'Không có sự kiện lịch nào trong phạm vi này.',
+        'review.past14': '14 ngày qua',
+        'review.past14Desc': 'Xem lại lịch của bạn trong hai tuần qua. Bạn có bỏ lỡ điều gì không? Có cuộc hẹn nào đã hoàn thành cần hành động theo dõi không?',
+        'review.upcoming14': '7 ngày tới',
+        'review.upcoming14Desc': 'Xem tuần tới. Bạn cần chuẩn bị gì? Ghi lại các hành động tiếp theo mới.',
+        'review.calendarTasks': 'Nhiệm vụ Mindwtr (7 ngày tới)',
+        'review.calendarTasksEmpty': 'Không có nhiệm vụ đã lên lịch/đến hạn trong phạm vi này.',
+        'review.waitingStep': 'Đang chờ',
+        'review.waitingStepDesc': 'Theo dõi các nhiệm vụ đã giao.',
+        'review.waitingHint': 'Xem lại các mục này. Bạn đã nhận được những gì bạn đang chờ chưa? Bạn có cần gửi lời nhắc không?',
+        'review.waitingEmpty': 'Không có gì trong Đang chờ.',
+        'review.contextsStepDesc': 'Xem lại các ngữ cảnh của bạn và đảm bảo mỗi ngữ cảnh có các hành động tiếp theo rõ ràng.',
+        'review.contextsEmpty': 'Không có ngữ cảnh nào có nhiệm vụ đang hoạt động.',
+        'review.aiStep': 'Phân tích AI',
+        'review.aiStepDesc': 'Làm nổi bật các mục cũ và gợi ý dọn dẹp.',
+        'review.aiTitle': 'Xem lại AI',
+        'review.aiRun': 'Chạy phân tích',
+        'review.aiRunning': 'Đang phân tích...',
+        'review.aiEmpty': 'Không tìm thấy mục nào cũ.',
+        'review.aiApply': 'Áp dụng đã chọn',
+        'review.aiAction.someday': 'Chuyển sang Tương lai',
+        'review.aiAction.archive': 'Lưu trữ',
+        'review.aiAction.breakdown': 'Cần phân tích',
+        'review.aiAction.keep': 'Giữ lại',
+        'review.notDueYet': 'Chưa đến hạn',
+        'review.projectsStep': 'Xem lại Dự án',
+        'review.projectsStepDesc': 'Đảm bảo mọi dự án đang hoạt động đều có hành động tiếp theo.',
+        'review.projectsHint': 'Xem lại từng dự án. Nó có ít nhất một Hành động tiếp theo cụ thể không? Nếu không, thêm một ngay bây giờ. Đánh dấu các dự án đã hoàn thành là xong.',
+        'review.hasNextAction': 'Có hành động tiếp theo',
+        'review.needsAction': 'Cần hành động',
+        'review.noActiveTasks': 'Không có nhiệm vụ đang hoạt động',
+        'review.somedayStep': 'Tương lai/Có thể',
+        'review.somedayStepDesc': 'Xem lại các dự án bạn có thể muốn bắt đầu.',
+        'review.somedayHint': 'Xem lại danh sách Tương lai/Có thể của bạn. Có điều gì bạn muốn kích hoạt ngay bây giờ không? Hoặc xóa?',
+        'review.listEmpty': 'Danh sách trống.',
+        'review.allDone': 'Tất cả xong!',
+        'review.allDoneDesc': 'Bạn đã sẵn sàng cho tuần tới.',
+        'review.complete': 'Xem lại hoàn tất!',
+        'review.completeDesc': 'Bạn đã làm rõ các đầu vào, cập nhật danh sách và sẵn sàng làm việc.',
+        'review.finish': 'Hoàn tất',
+        'review.next': 'Tiếp theo',
+        'review.step': 'Bước',
+        'review.of': 'trong',
+        'review.back': 'Quay lại',
+        'review.nextStepBtn': 'Bước tiếp theo',
+        'review.stuckQuestion': 'Mục này có bị kẹt không?',
+        'review.stuckPrompt': 'Thử viết lại để có thể thực hiện hơn, hoặc xóa/ủ nó nếu không còn liên quan.',
+        'review.notSet': 'Chưa đặt',
+        'review.description': 'Mô tả',
+        'review.startTime': 'Hoãn đến',
+        'review.deadline': 'Hạn chót',
+        'review.contexts': 'Ngữ cảnh',
+        'review.markDone': 'Đánh dấu hoàn thành',
+        'review.markReviewed': 'Đánh dấu đã xem lại',
+        'review.markReviewedDone': 'Đã đánh dấu xem lại',
+        'review.noTasks': 'Bạn đã xem hết. Không có gì để xem lại ngay bây giờ.',
+        'review.addTaskPlaceholder': 'Nhập tiêu đề nhiệm vụ',
+        'review.activeTasks': 'nhiệm vụ đang hoạt động',
+        'review.moreItems': 'mục nữa',
+        'review.noArea': 'Không có khu vực',
+        'review.singleActions': 'Hành động đơn lẻ',
+
+        // Processing
+        'process.title': 'Xử lý mục',
+        'process.refineTitle': 'Tinh chỉnh nhiệm vụ',
+        'process.refineDesc': 'Làm rõ đây là gì trước khi quyết định làm gì với nó.',
+        'process.refineNext': 'Tiếp theo',
+        'process.refineDelete': 'Xóa',
+        'process.actionable': 'Đây có thể thực hiện không?',
+        'process.actionableDesc': 'Bạn có thể thực hiện hành động vật lý trên cái này không?',
+        'process.yesActionable': '✅ Có, có thể thực hiện',
+        'process.moreThanOneStep': 'Nhiều hơn một bước?',
+        'process.moreThanOneStepDesc': 'Điều này có cần nhiều hành động và thuộc về một dự án không?',
+        'process.moreThanOneStepYes': 'Có, tạo thành dự án',
+        'process.moreThanOneStepNo': 'Không, hành động đơn lẻ',
+        'process.trash': 'Thùng rác',
+        'process.someday': 'Tương lai',
+        'process.reference': 'Tham khảo',
+        'process.later': 'Sau này',
+        'process.laterHint': 'Đặt ngày bắt đầu và chuyển sang Tiếp theo.',
+        'process.laterStartRequired': 'Chọn ngày bắt đầu cho Sau này.',
+        'process.twoMin': '⏱️ Liệu có mất dưới 2 phút không?',
+        'process.twoMinDesc': 'Nếu có, hãy làm ngay!',
+        'process.doneIt': 'Xong rồi!',
+        'process.takesLonger': 'Mất lâu hơn',
+        'process.nextStep': 'Tiếp theo là gì?',
+        'process.nextStepDesc': 'Bạn tự làm hay giao cho người khác?',
+        'process.doIt': '📋 Tôi sẽ làm',
+        'process.delegate': 'Giao cho người khác',
+        'process.delegateTitle': 'Giao',
+        'process.delegateDesc': 'Tùy chọn ghi chú người nhận và ngày theo dõi.',
+        'process.delegateWhoLabel': 'Ai? (tùy chọn)',
+        'process.delegateWhoPlaceholder': 'vd. Alex',
+        'process.delegateFollowUpLabel': 'Ngày theo dõi (tùy chọn)',
+        'process.delegateSendRequest': 'Gửi yêu cầu...',
+        'process.delegateMoveToWaiting': 'Chuyển sang Đang chờ',
+        'process.context': 'Bạn sẽ làm điều này ở đâu?',
+        'process.contextDesc': 'Thêm ngữ cảnh để tìm lại sau',
+        'process.selectMultipleHint': '(Chọn nhiều hoặc không chọn)',
+        'process.selectedLabel': 'Đã chọn:',
+        'process.ifNotActionable': 'Nếu không thể thực hiện:',
+        'process.waitingFor': 'Bạn đang chờ ai/điều gì?',
+        'process.waitingForDesc': 'Thêm ghi chú để nhớ bạn đang chờ gì',
+        'process.waitingPlaceholder': 'vd. Đang chờ John xem xét tài liệu...',
+        'process.next': 'Tiếp theo',
+        'process.noContext': 'Không có ngữ cảnh',
+        'process.project': 'Gán vào dự án?',
+        'process.projectDesc': 'Tùy chọn - liên kết nhiệm vụ này với dự án',
+        'process.noProject': 'Không cần dự án',
+        'process.makeProject': 'Tạo thành dự án',
+        'process.useExistingProject': 'Dùng dự án hiện có',
+        'process.nextAction': 'Hành động tiếp theo',
+        'process.createProject': 'Tạo dự án & thêm hành động tiếp theo',
+        'process.nextActionRequired': 'Thêm hành động tiếp theo trước khi tạo dự án.',
+        'process.skip': 'Bỏ qua & Thêm vào Tiếp theo',
+        'process.addContext': 'Thêm',
+        'process.newContextPlaceholder': 'Ngữ cảnh mới...',
+        'process.remaining': 'mục còn lại',
+        'process.btn': 'Xử lý Hộp thư đến',
+        'process.modeGuided': 'Có hướng dẫn',
+        'process.modeQuick': 'Nhanh',
+        'process.quickDesc': 'Xem lại chi tiết, chọn kết quả và chuyển ngay sang mục tiếp theo.',
+        'process.quickApplyHint': 'Áp dụng các lựa chọn này và tiếp tục sang mục hộp thư đến tiếp theo.',
+        'process.quickMoveHint': 'Di chuyển mục này và tiếp tục sang mục hộp thư đến tiếp theo.',
+
+        // Settings
+        'settings.title': 'Cài đặt',
+        'settings.subtitle': 'Tùy chỉnh trải nghiệm Mindwtr của bạn',
+        'settings.appearance': 'Giao diện',
+        'settings.general': 'Chung',
+        'settings.system': 'Hệ thống',
+        'settings.systemDesc': 'Theo cài đặt hệ thống',
+        'settings.light': 'Sáng',
+        'settings.lightDesc': 'Chế độ sáng',
+        'settings.dark': 'Tối',
+        'settings.darkDesc': 'Chế độ tối',
+        'settings.theme': 'Chủ đề',
+        'settings.eink': 'Màn hình mực điện tử',
+        'settings.nord': 'Nord',
+        'settings.sepia': 'Nâu đỏ',
+        'settings.oled': 'OLED / Đêm khuya',
+        'settings.material3Light': 'Material 3 (Sáng)',
+        'settings.material3Dark': 'Material 3 (Tối)',
+        'settings.language': 'Ngôn ngữ',
+        'settings.weekStart': 'Tuần bắt đầu vào',
+        'settings.weekStartSunday': 'Chủ nhật',
+        'settings.weekStartMonday': 'Thứ hai',
+        'settings.weekStartSaturday': 'Thứ bảy',
+        'settings.dateFormat': 'Định dạng ngày',
+        'settings.dateFormatSystem': 'Mặc định hệ thống',
+        'settings.dateFormatDmy': 'Ngày / Tháng / Năm',
+        'settings.dateFormatMdy': 'Tháng / Ngày / Năm',
+        'settings.dateFormatYmd': 'Năm / Tháng / Ngày',
+        'settings.calendarSystem': 'Hệ thống lịch',
+        'settings.calendarSystemGregorian': 'Dương lịch',
+        'settings.calendarSystemJalali': 'Jalali (Dương lịch Hijri)',
+        'settings.timeFormat': 'Định dạng giờ',
+        'settings.timeFormatSystem': 'Mặc định hệ thống',
+        'settings.timeFormat12h': '12 giờ',
+        'settings.timeFormat24h': '24 giờ',
+        'settings.keybindingsDesc': 'Chọn kiểu phím tắt máy tính ưa thích của bạn.',
+        'settings.closeBehavior': 'Hành vi đóng cửa sổ',
+        'settings.closeBehaviorDesc': 'Chọn điều gì xảy ra khi bạn đóng cửa sổ.',
+        'settings.closeBehaviorAsk': 'Hỏi mỗi lần',
+        'settings.closeBehaviorTray': 'Tiếp tục chạy trong khay hệ thống',
+        'settings.closeBehaviorQuit': 'Thoát ứng dụng',
+        'settings.closeBehaviorPromptTitle': 'Đóng Mindwtr?',
+        'settings.closeBehaviorPromptBody': 'Bạn muốn Mindwtr tiếp tục chạy trong khay hay thoát hoàn toàn?',
+        'settings.closeBehaviorRemember': 'Đừng hỏi lại',
+        'settings.about': 'Giới thiệu',
+        'settings.version': 'Phiên bản',
+        'settings.platform': 'Nền tảng',
+        'settings.diagnostics': 'Chẩn đoán',
+        'settings.diagnosticsDesc': 'Giúp khắc phục sự cố',
+        'settings.analyticsHeartbeat': 'Từ chối chẩn đoán',
+        'settings.analyticsHeartbeatDesc': 'Chẩn đoán là ẩn danh và giúp cải thiện Mindwtr. Không có văn bản nhiệm vụ, ghi chú hay dữ liệu cá nhân. Bật để ngừng gửi.',
+        'settings.analyticsHeartbeatDisableTitle': 'Từ chối chẩn đoán?',
+        'settings.analyticsHeartbeatDisableDesc': 'Bật để ngừng chẩn đoán ẩn danh. Không có văn bản nhiệm vụ, ghi chú hay dữ liệu cá nhân được gửi.',
+        'settings.analyticsHeartbeatDisableConfirm': 'Từ chối',
+        'settings.analyticsHeartbeatKeepEnabled': 'Tiếp tục gửi',
+        'settings.debugLogging': 'Ghi nhật ký gỡ lỗi',
+        'settings.debugLoggingDesc': 'Ghi lỗi cục bộ để chia sẻ với bộ phận hỗ trợ',
+        'settings.logFile': 'Tệp nhật ký',
+        'settings.clearLog': 'Xóa nhật ký',
+        'settings.shareLog': 'Chia sẻ nhật ký',
+        'settings.logMissing': 'Chưa tìm thấy tệp nhật ký. Bật ghi nhật ký và tái hiện sự cố trước.',
+        'settings.logCleared': 'Đã xóa tệp nhật ký.',
+        'settings.shareUnavailable': 'Chia sẻ không khả dụng trên thiết bị này.',
+        'settings.useSystem': 'Dùng chủ đề hệ thống',
+        'settings.followDevice': 'Theo cài đặt giao diện thiết bị',
+        'settings.darkMode': 'Chế độ tối',
+        'settings.darkEnabled': 'Chủ đề tối đã bật',
+        'settings.lightEnabled': 'Chủ đề sáng đã bật',
+        'settings.material3Theme': 'Chủ đề Material 3',
+        'settings.material3ThemeDesc': 'Dùng token màu Material 3 trên Android',
+        'settings.selectLang': 'Chọn ngôn ngữ ưa thích của bạn',
+        'settings.privacy': 'Quyền riêng tư',
+        'settings.mobile.appLock': 'Khóa ứng dụng',
+        'settings.mobile.appLockDesc': 'Yêu cầu khóa thiết bị khi mở Mindwtr hoặc quay lại ứng dụng. Điều này bảo vệ chế độ xem ứng dụng, không phải cơ sở dữ liệu trên thiết bị.',
+        'appLock.title': 'Mindwtr đã bị khóa',
+        'appLock.description': 'Dùng khóa thiết bị để mở Mindwtr. Điều này bảo vệ chế độ xem ứng dụng, không phải cơ sở dữ liệu trên thiết bị.',
+        'appLock.prompt': 'Mở khóa Mindwtr',
+        'appLock.enablePrompt': 'Bật khóa ứng dụng Mindwtr',
+        'appLock.unlock': 'Mở khóa',
+        'appLock.authenticating': 'Đang xác thực...',
+        'appLock.useDevicePasscode': 'Dùng mã PIN thiết bị',
+        'appLock.unavailable': 'Thiết lập mã PIN hoặc sinh trắc học để dùng khóa ứng dụng.',
+        'appLock.cancelled': 'Mindwtr vẫn bị khóa.',
+        'appLock.failed': 'Xác thực thất bại. Thử lại.',
+        'settings.data': 'Dữ liệu',
+        'settings.menuDesc.general': 'Giao diện, ngôn ngữ, hiển thị',
+        'settings.menuDesc.gtd': 'Ghi lại, xem lại, trình chỉnh sửa nhiệm vụ',
+        'settings.menuDesc.manage': 'Khu vực, ngữ cảnh, thẻ',
+        'settings.menuDesc.notifications': 'Nhắc nhở, bản tóm tắt hàng ngày',
+        'settings.menuDesc.sync': 'WebDAV, Dropbox, iCloud',
+        'settings.menuDesc.data': 'Sao lưu, khôi phục, nhập',
+        'settings.menuDesc.advanced': 'Trợ lý AI, lịch',
+        'settings.menuDesc.about': 'Phiên bản, giấy phép, liên kết',
+        'settings.menuDesc.ai': 'Mô hình, nhà cung cấp, giọng nói',
+        'settings.menuDesc.calendar': 'Lịch và sự kiện bên ngoài',
+        'settings.dataSync': 'Dữ liệu & Đồng bộ',
+        'settings.dataDesc': 'Nhập hoặc xuất dữ liệu để đồng bộ với các thiết bị khác.',
+        'settings.calendar': 'Lịch bên ngoài',
+        'settings.calendarDesc': 'Kết nối lịch bên ngoài (URL đăng ký ICS hoặc tệp .ics cục bộ).',
+        'settings.deviceCalendars': 'Lịch thiết bị',
+        'settings.deviceCalendarsDesc': 'Đọc sự kiện từ lịch đã đồng bộ trên thiết bị này (DAVx5, iCloud, Outlook, v.v.).',
+        'settings.calendarAccessDenied': 'Quyền truy cập lịch bị từ chối. Bật trong cài đặt hệ thống, sau đó làm mới.',
+        'settings.calendarAccessRequired': 'Cần quyền truy cập lịch để đọc sự kiện thiết bị.',
+        'settings.grantCalendarAccess': 'Cấp quyền truy cập',
+        'settings.noDeviceCalendars': 'Không tìm thấy lịch nào trên thiết bị.',
+        'settings.deviceCalendar': 'Lịch thiết bị',
+        'settings.externalCalendars': 'Lịch bên ngoài',
+        'settings.externalCalendarName': 'Tên',
+        'settings.externalCalendarUrl': 'URL ICS',
+        'settings.externalCalendarUrlPlaceholder': 'https://example.com/calendar.ics',
+        'settings.externalCalendarAdd': 'Thêm lịch',
+        'settings.externalCalendarRemove': 'Xóa',
+        'settings.syncBackend': 'Backend đồng bộ',
+        'settings.syncBackendFile': 'Tệp',
+        'settings.syncBackendOff': 'Tắt',
+        'settings.syncBackendWebdav': 'WebDAV',
+        'settings.syncBackendCloud': 'Tự lưu trữ',
+        'settings.syncBackendChoiceHint': 'Chọn một phương thức đồng bộ. Tắt đồng bộ để giữ dữ liệu cục bộ trên thiết bị này.',
+        'settings.syncBackendGroupCloud': 'Đồng bộ đám mây',
+        'settings.syncBackendGroupCloudDesc': 'Đồng bộ tài khoản dễ dàng do Mindwtr quản lý.',
+        'settings.syncBackendGroupFile': 'Đồng bộ thư mục / tệp',
+        'settings.syncBackendGroupFileDesc': 'Mang thư mục đồng bộ cục bộ của riêng bạn.',
+        'settings.syncBackendGroupAdvanced': 'Nâng cao / Máy chủ tùy chỉnh',
+        'settings.syncBackendGroupAdvancedDesc': 'Dùng WebDAV hoặc điểm cuối Mindwtr Cloud của riêng bạn.',
+        'settings.syncOff': 'Đồng bộ đang tắt',
+        'settings.syncOffDesc': 'Bật lại đồng bộ bất cứ lúc nào từ màn hình này.',
+        'settings.syncPreferences': 'Tùy chọn đồng bộ cài đặt',
+        'settings.syncPreferencesDesc': 'Chọn cài đặt nào được đồng bộ trên các thiết bị.',
+        'settings.syncPreferenceAppearance': 'Giao diện (chủ đề, tuổi nhiệm vụ, truy cập nhanh)',
+        'settings.syncPreferenceLanguage': 'Ngôn ngữ & định dạng ngày/giờ',
+        'settings.syncPreferenceGtd': 'Mặc định lên lịch',
+        'settings.syncPreferenceExternalCalendars': 'URL lịch bên ngoài',
+        'settings.syncPreferenceAi': 'Cài đặt AI (mô hình, nhà cung cấp)',
+        'settings.syncPreferenceAiHint': 'Khóa API và đường dẫn mô hình cục bộ không bao giờ được đồng bộ.',
+        'settings.lastSync': 'Đồng bộ lần cuối',
+        'settings.lastSyncNever': 'Chưa bao giờ',
+        'settings.lastSyncSuccess': 'Đồng bộ hoàn thành',
+        'settings.lastSyncError': 'Đồng bộ thất bại',
+        'settings.lastSyncConflict': 'Đã giải quyết xung đột',
+        'settings.syncStatusFresh': 'Đã đồng bộ',
+        'settings.syncStatusStale': 'Cũ',
+        'settings.syncStatusOld': 'Quá cũ',
+        'settings.syncStatusNever': 'Chưa đồng bộ',
+        'settings.syncConflictNotice': 'Xung đột đồng bộ được giải quyết bằng ghi cuối cùng. Mở Cài đặt → Đồng bộ để xem chi tiết.',
+        'settings.queuedSyncFailed': 'Đồng bộ hàng đợi thất bại.',
+        'settings.attachmentSyncRetryWarning': 'Đồng bộ tệp đính kèm vẫn thất bại. Các tệp sẽ thử lại trong nền.',
+        'settings.syncBadgeSyncing': 'Đang đồng bộ',
+        'settings.syncBadgeHealthy': 'Đồng bộ ổn định',
+        'settings.syncBadgeWarning': 'Đồng bộ cần chú ý',
+        'settings.syncingAccessibility': 'Đang đồng bộ. Nhấn để mở cài đặt để biết chi tiết đồng bộ.',
+        'settings.syncIssueAccessibility': 'Đồng bộ cần chú ý. Nhấn để mở cài đặt để biết chi tiết đồng bộ.',
+        'settings.syncFailureOffline': 'Kiểm tra kết nối internet và thử lại.',
+        'settings.syncFailureAuth': 'Xác thực lại hoặc xem xét thông tin đăng nhập đồng bộ trong Cài đặt → Đồng bộ.',
+        'settings.syncFailurePermission': 'Chọn lại tệp hoặc thư mục đồng bộ, hoặc cấp quyền truy cập lại trong Cài đặt → Đồng bộ.',
+        'settings.syncFailureRateLimited': 'Backend đồng bộ đang giới hạn tốc độ yêu cầu. Chờ một lúc và thử lại.',
+        'settings.syncFailureMisconfigured': 'Hoàn tất cấu hình backend đồng bộ đã chọn trong Cài đặt → Đồng bộ.',
+        'settings.syncFailureConflict': 'Thiết bị hoặc backend khác báo cáo xung đột đồng bộ. Thử lại sau khi cả hai phía hoàn thành đồng bộ.',
+        'settings.syncFailureGeneric': 'Xem xét Cài đặt → Đồng bộ và thử lại.',
+        'settings.syncSkippedOffline': 'Không có kết nối internet. Đã bỏ qua đồng bộ.',
+        'settings.syncQueuedBody': 'Các thay đổi cục bộ đã đến trong khi đồng bộ. Lần thử lại đã được xếp hàng tự động.',
+        'settings.syncCompleted': 'Đồng bộ hoàn thành!',
+        'settings.syncCompletedWithConflicts': 'Đồng bộ hoàn thành với {count} xung đột (đã giải quyết tự động).',
+        'settings.syncClockSkewWarning': 'Phát hiện độ lệch đồng hồ thiết bị lớn ({skew}). Kiểm tra cài đặt thời gian trên mỗi thiết bị.',
+        'settings.syncAdjustedTimestamps': 'Đã điều chỉnh {count} dấu thời gian trong tương lai trong khi đồng bộ.',
+        'settings.webdavUrl': 'URL WebDAV',
+        'settings.webdavUrlPlaceholder': 'https://example.com/remote.php/dav/files/<tên người dùng>/mindwtr',
+        'settings.webdavUsername': 'Tên người dùng',
+        'settings.webdavUsernamePlaceholder': 'tên người dùng',
+        'settings.webdavPassword': 'Mật khẩu',
+        'settings.webdavSave': 'Lưu WebDAV',
+        'settings.webdavHint': 'Trỏ đến thư mục. Mindwtr lưu data.json bên trong đó.',
+        'settings.allowInsecureHttp': 'Cho phép kết nối không an toàn (HTTP)',
+        'settings.allowInsecureHttpHint': 'Chỉ HTTP mạng cục bộ hoặc riêng tư được phép. Dữ liệu sẽ được gửi không mã hóa.',
+        'settings.cleartextSyncWarningWebdav': 'Đồng bộ WebDAV đang dùng HTTP. Chỉ cho phép điểm cuối mạng cục bộ hoặc riêng tư; dữ liệu không được mã hóa.',
+        'settings.cleartextSyncWarningCloud': 'Đồng bộ tự lưu trữ đang dùng HTTP. Chỉ cho phép điểm cuối mạng cục bộ hoặc riêng tư; dữ liệu không được mã hóa.',
+        'settings.webdavBrowserStorageWarning': 'Mật khẩu WebDAV được lưu trong bộ nhớ trình duyệt. Chỉ dùng trên thiết bị đáng tin cậy.',
+        'settings.cloudUrl': 'URL tự lưu trữ',
+        'settings.cloudUrlPlaceholder': 'https://example.com',
+        'settings.cloudToken': 'Token truy cập',
+        'settings.cloudSave': 'Lưu tự lưu trữ',
+        'settings.cloudHint': 'Dùng URL điểm cuối tự lưu trữ của bạn.',
+        'settings.cloudBaseUrlHint': 'Dùng URL gốc. Mindwtr tự động thêm /v1/data.',
+        'settings.invalidUrlHttp': 'URL không hợp lệ. Dùng http/https.',
+        'settings.testConnection': 'Kiểm tra kết nối',
+        'settings.importTodoistSummary': 'Đã nhập {{taskCount}} nhiệm vụ vào {{projectCount}} dự án.',
+        'settings.importDgtSummary': 'Đã nhập {{taskCount}} nhiệm vụ, {{projectCount}} dự án và {{areaCount}} khu vực.',
+        'settings.importOmniFocusSummary': 'Đã nhập {{taskCount}} nhiệm vụ và {{projectCount}} dự án.',
+        'settings.webdavTestHint': 'Xác minh URL và thông tin đăng nhập mà không đồng bộ dữ liệu',
+        'settings.webdavTestAccessibility': 'Kiểm tra kết nối WebDAV',
+        'settings.cloudTestHint': 'Xác minh URL và token mà không đồng bộ dữ liệu',
+        'settings.cloudTestAccessibility': 'Kiểm tra kết nối tự lưu trữ',
+        'settings.dropboxTestHint': 'Xác minh token Dropbox và quyền truy cập tài khoản.',
+        'settings.dropboxTestAccessibility': 'Kiểm tra kết nối Dropbox',
+        'settings.importData': 'Nhập dữ liệu',
+        'settings.exportData': 'Xuất dữ liệu',
+        'settings.importSuccess': 'Nhập dữ liệu thành công!',
+        'settings.exportSuccess': 'Xuất dữ liệu thành công!',
+        'settings.notifications': 'Thông báo',
+        'settings.notificationsDesc': 'Bật nhắc nhở cho các nhiệm vụ có ngày bắt đầu hoặc đến hạn.',
+        'settings.notificationsEnable': 'Nhắc nhở nhiệm vụ',
+        'settings.notificationsDisabled': 'Thông báo đã tắt',
+        'settings.notificationsDisabledMessage': 'Mindwtr không thể lên lịch nhắc nhở cho đến khi quyền thông báo được khôi phục.',
+        'settings.startDateNotifications': 'Nhắc nhở ngày bắt đầu',
+        'settings.startDateNotificationsDesc': 'Thông báo khi ngày/giờ bắt đầu nhiệm vụ đến.',
+        'settings.dueDateNotifications': 'Nhắc nhở ngày đến hạn',
+        'settings.dueDateNotificationsDesc': 'Thông báo khi ngày/giờ đến hạn nhiệm vụ đến.',
+        'settings.reviewAtNotifications': 'Nhắc nhở ngày xem lại',
+        'settings.reviewAtNotificationsDesc': 'Thông báo khi ngày xem lại (nhắc nhở) đến hạn.',
+        'settings.notificationsDevHint': 'Thông báo di động yêu cầu bản dựng phát triển (không phải Expo Go).',
+        'settings.ai': 'Trợ lý AI',
+        'settings.advanced': 'Nâng cao',
+        'settings.aiDesc': 'Trợ giúp tùy chọn để làm rõ và phân tích nhiệm vụ.',
+        'settings.aiEnable': 'Bật trợ lý AI',
+        'settings.aiProvider': 'Nhà cung cấp',
+        'settings.aiModel': 'Mô hình',
+        'settings.aiBaseUrl': 'URL gốc tương thích OpenAI tùy chỉnh',
+        'settings.aiBaseUrlHint': 'Để trống cho OpenAI chính thức. Đặt cho API tương thích OpenAI cục bộ hoặc bên thứ ba như llama.cpp, Ollama, LM Studio, GLM hoặc vLLM.',
+        'settings.aiBaseUrlPlaceholder': 'http://localhost:11434/v1',
+        'settings.aiExtraBodyParams': 'Tham số yêu cầu bổ sung',
+        'settings.aiExtraBodyParamsDesc': 'Đối tượng JSON tùy chọn được hợp nhất vào các yêu cầu tương thích OpenAI.',
+        'settings.aiExtraBodyParamsHint': 'Các trường cốt lõi như model, messages và response_format được bảo vệ. Ví dụ: { "thinking": { "type": "disabled" } }',
+        'settings.aiExtraBodyParamsInvalid': 'Nhập đối tượng JSON hợp lệ.',
+        'settings.aiExtraBodyParamsSave': 'Lưu tham số',
+        'settings.aiApiKey': 'Khóa API',
+        'settings.aiApiKeyHint': 'Được lưu cục bộ trên thiết bị này. Không bao giờ đồng bộ. OpenAI chính thức yêu cầu khóa. Các điểm cuối tương thích OpenAI tùy chỉnh cũng có thể yêu cầu; để trống chỉ khi điểm cuối của bạn cho phép yêu cầu không xác thực.',
+        'settings.aiApiKeyPlaceholder': 'Dán khóa API của bạn',
+        'settings.speechTitle': 'Chuyển giọng nói thành văn bản',
+        'settings.speechDesc': 'Phiên âm các ghi âm giọng nói và ánh xạ chúng vào các trường nhiệm vụ.',
+        'settings.speechEnable': 'Bật chuyển giọng nói thành văn bản',
+        'settings.speechProvider': 'Nhà cung cấp giọng nói',
+        'settings.speechProviderOffline': 'Trên thiết bị (Whisper)',
+        'settings.speechModel': 'Mô hình giọng nói',
+        'settings.speechOfflineModel': 'Mô hình ngoại tuyến',
+        'settings.speechOfflineModelDesc': 'Tải xuống một lần để phiên âm hoàn toàn ngoại tuyến.',
+        'settings.speechOfflineReady': 'Mô hình đã tải xuống',
+        'settings.speechOfflineNotDownloaded': 'Mô hình chưa tải xuống',
+        'settings.speechOfflineDownload': 'Tải xuống',
+        'settings.speechOfflineDownloadSuccess': 'Tải xuống hoàn thành',
+        'settings.speechOfflineDelete': 'Xóa',
+        'settings.speechOfflineDownloadError': 'Tải xuống mô hình ngoại tuyến thất bại',
+        'settings.speechOfflineDeleteError': 'Xóa mô hình ngoại tuyến thất bại',
+        'settings.speechOfflineDeleteErrorBody': 'Vui lòng thử lại.',
+        'settings.speechLanguage': 'Ngôn ngữ âm thanh',
+        'settings.speechLanguageHint': 'Dùng tên hoặc mã ngôn ngữ, hoặc để trống để tự động phát hiện.',
+        'settings.speechLanguageAuto': 'Tự động (phát hiện ngôn ngữ)',
+        'settings.speechMode': 'Chế độ xử lý',
+        'settings.speechModeHint': 'Phân tích thông minh trích xuất ngày và trường; chỉ phiên âm chỉ phiên âm.',
+        'settings.speechModeSmart': 'Phân tích thông minh',
+        'settings.speechModeTranscript': 'Chỉ phiên âm',
+        'settings.speechFieldStrategy': 'Ánh xạ trường',
+        'settings.speechFieldStrategyHint': 'Chọn nơi bản phiên âm sẽ đến theo mặc định.',
+        'settings.speechFieldSmart': 'Thông minh',
+        'settings.speechFieldTitle': 'Tiêu đề',
+        'settings.speechFieldDescription': 'Mô tả',
+        'settings.aiReasoning': 'Nỗ lực suy luận',
+        'settings.aiReasoningHint': 'Dùng cho các mô hình GPT-5.',
+        'settings.aiReasoningHintFoss': 'Dùng cho các mô hình được hỗ trợ.',
+        'settings.aiProviderOpenAI': 'OpenAI',
+        'settings.aiProviderGemini': 'Gemini',
+        'settings.aiProviderAnthropic': 'Anthropic (Claude)',
+        'settings.aiEffortLow': 'Thấp',
+        'settings.aiEffortMedium': 'Trung bình',
+        'settings.aiEffortHigh': 'Cao',
+        'settings.aiCopilotModel': 'Mô hình Copilot',
+        'settings.aiCopilotHint': 'Dùng cho gợi ý tự động hoàn thành nhanh.',
+        'settings.aiThinkingBudget': 'Ngân sách suy nghĩ',
+        'settings.aiThinkingHint': 'Chỉ Claude/Gemini. 0 vô hiệu hóa suy nghĩ mở rộng.',
+        'settings.aiThinkingEnable': 'Bật suy nghĩ',
+        'settings.aiThinkingEnableDesc': 'Dùng suy luận mở rộng cho các nhiệm vụ phức tạp.',
+        'settings.aiThinkingOff': 'Tắt',
+        'settings.aiThinkingLow': 'Thấp',
+        'settings.aiThinkingMedium': 'Trung bình',
+        'settings.aiThinkingHigh': 'Cao',
+        'ai.disabledTitle': 'Trợ lý AI đang tắt',
+        'ai.disabledBody': 'Bật AI trong Cài đặt để dùng tính năng này.',
+        'ai.missingKeyTitle': 'Cần khóa API',
+        'ai.missingKeyBody': 'Thêm khóa API trong Cài đặt → Trợ lý AI.',
+        'ai.errorTitle': 'Yêu cầu AI thất bại',
+        'ai.errorBody': 'Vui lòng thử lại.',
+        'ai.applySuggestion': 'Dùng gợi ý',
+        'ai.addSteps': 'Thêm các bước',
+        'ai.working': 'Đang xử lý...',
+        'ai.breakdownTitle': 'Các bước được gợi ý',
+        'copilot.suggested': 'Được gợi ý',
+        'copilot.applyHint': 'Nhấn để áp dụng',
+        'copilot.applied': 'Đã áp dụng',
+        'settings.dailyDigest': 'Bản tóm tắt hàng ngày',
+        'settings.dailyDigestDesc': 'Nhận bản tóm tắt buổi sáng và nhắc nhở xem lại buổi tối.',
+        'settings.dailyDigestMorning': 'Bản tóm tắt buổi sáng',
+        'settings.dailyDigestMorningTime': 'Giờ buổi sáng',
+        'settings.dailyDigestEvening': 'Xem lại buổi tối',
+        'settings.dailyDigestEveningTime': 'Giờ buổi tối',
+        'settings.gtd': 'GTD',
+        'settings.gtdDesc': 'Mindwtr hoạt động ngay khi dùng. Chỉ điều chỉnh các mặc định này khi bạn muốn quy trình GTD khác.',
+        'settings.inboxProcessing': 'Xử lý hộp thư đến',
+        'settings.inboxProcessingDesc': 'Điều chỉnh quy trình hộp thư đến theo cách bạn muốn xử lý nhiệm vụ.',
+        'settings.inboxTwoMinuteFirst': 'Hỏi quy tắc 2 phút trước',
+        'settings.inboxTwoMinuteEnabled': 'Hiện phím tắt 2 phút',
+        'settings.inboxProjectFirst': 'Hỏi về dự án sớm',
+        'settings.inboxContextStepEnabled': 'Bao gồm phần Ngữ cảnh & Thẻ',
+        'settings.inboxScheduleEnabled': 'Cho phép lên lịch trong khi xử lý',
+        'settings.inboxReferenceEnabled': 'Đề xuất Tham khảo trong khi xử lý',
+        'settings.captureDefault': 'Phương thức ghi lại mặc định',
+        'settings.captureDefaultDesc': 'Chọn ghi lại nhanh bắt đầu bằng văn bản hay âm thanh.',
+        'settings.captureDefaultText': 'Văn bản',
+        'settings.captureDefaultAudio': 'Âm thanh',
+        'settings.captureSaveAudio': 'Lưu tệp đính kèm âm thanh',
+        'settings.captureSaveAudioDesc': 'Giữ tệp âm thanh đính kèm sau khi phiên âm.',
+        'settings.autoArchive': 'Tự động lưu trữ nhiệm vụ đã xong',
+        'settings.autoArchiveDesc': 'Chuyển các nhiệm vụ đã hoàn thành sang Đã lưu trữ sau số ngày đã chọn.',
+        'settings.autoArchiveDays': 'Lưu trữ sau',
+        'settings.autoArchiveNever': 'Không bao giờ (giữ trong Hoàn thành)',
+        'settings.focusTaskLimit': 'Giới hạn tập trung hôm nay',
+        'settings.focusTaskLimitDesc': 'Số nhiệm vụ tối đa bạn có thể đánh dấu sao cho tập trung hôm nay.',
+        'settings.defaultProjectFlowMode': 'Luồng dự án mặc định',
+        'settings.defaultProjectFlowModeDesc': 'Chỉ áp dụng khi tạo dự án mới. Các dự án hiện có giữ luồng hiện tại.',
+        'settings.projectFlowParallel': 'Song song',
+        'settings.projectFlowSequential': 'Tuần tự',
+        'settings.timeEstimatePresets': 'Cài đặt sẵn ước tính thời gian',
+        'settings.timeEstimatePresetsDesc': 'Chọn ước tính thời gian nào xuất hiện trong trình chỉnh sửa nhiệm vụ.',
+        'settings.timeEstimatePresetsDisabled': 'Bật Ước tính thời gian để tùy chỉnh cài đặt sẵn.',
+        'settings.enableTimeEstimates': 'Bật Ước tính thời gian',
+        'settings.features': 'Tính năng',
+        'settings.featuresDesc': 'Các tín hiệu tùy chọn bạn có thể bật khi cần.',
+        'settings.featurePriorities': 'Ưu tiên',
+        'settings.featurePrioritiesDesc': 'Hiện cờ ưu tiên trên nhiệm vụ.',
+        'settings.featureTimeEstimates': 'Ước tính thời gian',
+        'settings.featureTimeEstimatesDesc': 'Thêm ước tính thời lượng nhanh để chặn thời gian.',
+        'settings.featurePomodoro': 'Hẹn giờ Pomodoro',
+        'settings.featurePomodoroDesc': 'Bật bảng Pomodoro tùy chọn trong chế độ Tập trung.',
+        'settings.pomodoroCustomPreset': 'Cài đặt sẵn tùy chỉnh',
+        'settings.pomodoroCustomPresetDesc': 'Thêm một cài đặt sẵn tập trung/nghỉ thêm. Khớp với cài đặt sẵn tích hợp chỉ giữ các chip tích hợp.',
+        'settings.pomodoroFocusMinutes': 'Phút tập trung',
+        'settings.pomodoroBreakMinutes': 'Phút nghỉ',
+        'settings.pomodoroLinkTask': 'Liên kết hẹn giờ với nhiệm vụ',
+        'settings.pomodoroLinkTaskDesc': 'Hiện bộ chọn nhiệm vụ Hẹn giờ và hành động Hoàn thành nhiệm vụ trong Tập trung.',
+        'settings.resetToDefault': 'Đặt lại về mặc định',
+        'settings.taskEditorLayout': 'Bố cục trình chỉnh sửa nhiệm vụ',
+        'settings.taskEditorLayoutDesc': 'Chọn các trường xuất hiện khi chỉnh sửa nhiệm vụ, bao gồm Ưu tiên và Ước tính thời gian.',
+        'settings.taskEditorLayoutHint': 'Nhấn để bật/tắt trường. Kéo tay cầm để sắp xếp lại. Các trường ẩn hiển thị trong "Thêm" hoặc khi đã có nội dung.',
+        'settings.taskEditorDefaultOpen': 'Mở các phần theo mặc định',
+        'settings.weeklyReview': 'Xem lại hàng tuần',
+        'settings.weeklyReviewDesc': 'Nhận nhắc nhở xem lại hàng tuần vào giờ đã chọn.',
+        'settings.weeklyReviewDay': 'Ngày xem lại',
+        'settings.weeklyReviewTime': 'Giờ xem lại',
+        'settings.dailyReviewConfig': 'Quy trình xem lại hàng ngày',
+        'settings.dailyReviewConfigDesc': 'Cấu hình lần khởi động hàng ngày nhẹ nhàng.',
+        'settings.dailyReviewIncludeFocusStep': 'Bao gồm bước Tập trung',
+        'settings.dailyReviewIncludeFocusStepDesc': 'Hiện bước tùy chọn để đánh dấu các nhiệm vụ tập trung hôm nay.',
+        'settings.weeklyReviewConfig': 'Quy trình xem lại hàng tuần',
+        'settings.weeklyReviewConfigDesc': 'Cấu hình các bước xuất hiện trong xem lại hàng tuần.',
+        'settings.weeklyReviewIncludeContextsStep': 'Bao gồm bước Ngữ cảnh',
+        'settings.weeklyReviewIncludeContextsStepDesc': 'Hiện bước xem lại ngữ cảnh chuyên dụng giữa Đang chờ và Dự án.',
+        'widget.capture': 'Ghi lại nhanh',
+        'obsidian.taskNotesDetectedTitle': 'Chế độ TaskNotes đang hoạt động',
+        'obsidian.taskNotesDetectedBody': 'Mindwtr phát hiện frontmatter kiểu TaskNotes trong các tệp này, vì vậy các nhiệm vụ danh sách kiểm tra nội tuyến từ các ghi chú khác bị bỏ qua.',
+        'obsidian.taskNotesDetectedHint': 'Tìm trường trạng thái cộng với siêu dữ liệu TaskNotes như tags: [task], due, scheduled, contexts, projects, timeEstimate, recurrence hoặc completedDate.',
+        'obsidian.taskNotesDetectedMore': 'tệp khớp thêm',
+        'obsidian.showCompleted': 'Hiện đã hoàn thành',
+        'obsidian.hideCompleted': 'Ẩn đã hoàn thành',
+        'obsidian.completedHidden': 'Đã ẩn hoàn thành',
+        'obsidian.completedOnlyTitle': 'Chỉ ẩn các nhiệm vụ đã hoàn thành',
+        'obsidian.completedOnlyBody': 'Kho lưu trữ này hiện chỉ có các nhiệm vụ đã nhập hoàn thành. Bật Hiện đã hoàn thành để kiểm tra hoặc mở lại chúng.',
+
+        'digest.morningTitle': 'Bản tóm tắt buổi sáng',
+        'digest.morningBody': 'Mở Mindwtr để lên kế hoạch cho ngày của bạn.',
+        'digest.eveningTitle': 'Xem lại buổi tối',
+        'digest.eveningBody': 'Mở Mindwtr để xem lại và kết thúc.',
+        'digest.weeklyReviewTitle': 'Xem lại hàng tuần',
+        'digest.weeklyReviewBody': 'Mở Mindwtr để xem lại và đặt lại tuần của bạn.',
+        'digest.dueToday': 'Đến hạn hôm nay',
+        'digest.overdue': 'Quá hạn',
+        'digest.focus': 'Tập trung',
+        'digest.reviewDue': 'Xem lại đến hạn',
+        'digest.noItems': 'Không có mục khẩn cấp hôm nay.',
+
+        // Common
+        'common.tasks': 'nhiệm vụ',
+        'common.days': 'ngày',
+        'common.cancel': 'Hủy',
+        'common.discard': 'Bỏ',
+        'common.save': 'Lưu',
+        'common.delete': 'Xóa',
+        'common.rename': 'Đổi tên',
+        'common.name': 'Tên',
+        'settings.deleteNamed': 'Xóa "{{name}}"?',
+        'people.deleteConfirm': 'Xóa người này? Các nhiệm vụ được giao cho họ sẽ được giữ lại và chuyển sang chưa phân công.',
+        'settings.unassignedAreaColor': 'Màu chưa phân công',
+        'settings.unassignedAreaColorDesc': 'Dùng cho các nhiệm vụ và dự án không có khu vực.',
+        'contexts.empty': 'Không có ngữ cảnh',
+        'tags.title': 'Thẻ',
+        'areas.edit': 'Chỉnh sửa khu vực',
+        'common.edit': 'Chỉnh sửa',
+        'common.add': 'Thêm',
+        'common.open': 'Mở',
+        'common.ok': 'OK',
+        'common.share': 'Chia sẻ',
+        'common.all': 'Tất cả',
+        'common.search': 'Tìm kiếm...',
+        'common.noMatches': 'Không có kết quả',
+        'common.validationRequired': 'Vui lòng nhập giá trị.',
+        'common.none': 'Không có',
+        'common.notSet': 'Chưa đặt',
+        'common.back': 'Quay lại',
+        'common.close': 'Đóng',
+        'common.dismiss': 'Bỏ qua',
+        'common.play': 'Phát',
+        'common.pause': 'Tạm dừng',
+        'common.done': 'Xong',
+        'common.showCompleted': 'Hiện đã hoàn thành',
+        'common.hideCompleted': 'Ẩn đã hoàn thành',
+        'common.undo': 'Hoàn tác',
+        'common.skip': 'Bỏ qua',
+        'common.clear': 'Xóa',
+        'common.comingSoon': 'Sắp ra mắt',
+        'common.more': 'Thêm',
+        'common.less': 'Ít hơn',
+        'common.loading': 'Đang tải...',
+        'common.offline': 'Ngoại tuyến',
+        'common.notice': 'Thông báo',
+        'common.success': 'Thành công',
+        'errorBoundary.title': 'Đã xảy ra lỗi',
+        'errorBoundary.message': 'Ứng dụng gặp lỗi không mong đợi.',
+        'errorBoundary.retry': 'Thử lại',
+        'audio.loading': 'Đang tải âm thanh...',
+        'attachments.retryTranscription': 'Phiên âm lại',
+        'attachments.transcribing': 'Đang phiên âm...',
+        'attachments.transcriptionUnavailable': 'Chuyển giọng nói thành văn bản chưa sẵn sàng. Kiểm tra cài đặt AI và thử lại.',
+        'attachments.transcriptionFailed': 'Phiên âm thất bại. Vui lòng thử lại.',
+
+        // Bulk Actions
+        'bulk.select': 'Chọn',
+        'bulk.exitSelect': 'Xong',
+        'bulk.selected': 'đã chọn',
+        'bulk.moveTo': 'Chuyển đến',
+        'bulk.selectRange': 'Phạm vi',
+        'bulk.selectRangeActive': 'Chọn điểm cuối',
+        'bulk.addTag': 'Thêm thẻ',
+        'bulk.addContext': 'Thêm ngữ cảnh',
+        'bulk.removeContext': 'Xóa ngữ cảnh',
+        'bulk.tagPlaceholder': '#thẻ',
+        'bulk.contextPlaceholder': '@ngữ cảnh',
+        'bulk.delete': 'Xóa',
+        'bulk.confirmDeleteTitle': 'Xóa nhiệm vụ',
+        'bulk.confirmDeleteBody': 'Xóa các nhiệm vụ đã chọn?',
+
+        // Sorting
+        'sort.label': 'Sắp xếp',
+        'sort.default': 'Mặc định',
+        'sort.due': 'Ngày đến hạn',
+        'sort.start': 'Ngày bắt đầu',
+        'sort.review': 'Ngày xem lại',
+        'sort.title': 'Tiêu đề',
+        'sort.created': 'Đã tạo (cũ nhất)',
+        'sort.created-desc': 'Đã tạo (mới nhất)',
+
+        // Agenda
+        'agenda.title': 'Tập trung',
+        'agenda.active': 'mục đang hoạt động',
+        'agenda.top3Only': 'Chỉ hiện Top 3',
+        'agenda.top3Title': 'Top 3 hôm nay',
+        'agenda.futureStartsShownOne': '1 nhiệm vụ bắt đầu tương lai được hiển thị',
+        'agenda.futureStartsShownMany': '{count} nhiệm vụ bắt đầu tương lai được hiển thị',
+        'agenda.futureStartsHiddenOne': '1 nhiệm vụ bị ẩn (bắt đầu tương lai)',
+        'agenda.futureStartsHiddenMany': '{count} nhiệm vụ bị ẩn (bắt đầu tương lai)',
+        'agenda.showFutureStarts': 'Hiện',
+        'agenda.hideFutureStarts': 'Ẩn',
+        'agenda.showMore': 'Hiện thêm {{count}} nhiệm vụ...',
+        'agenda.zenMode': 'Chế độ Zen',
+        'agenda.zenHidden': 'Và thêm {{count}} nhiệm vụ bị ẩn...',
+        'agenda.todaysFocus': 'Tập trung hôm nay',
+        'agenda.focusHint': 'Nhấn ngôi sao trên bất kỳ nhiệm vụ nào bên dưới để thêm vào tập trung hôm nay (tối đa {{count}}).',
+        'agenda.addToFocus': 'Thêm vào tập trung hôm nay',
+        'agenda.removeFromFocus': 'Xóa khỏi tập trung',
+        'agenda.maxFocusItems': 'Tối đa {{count}} mục tập trung',
+        'projects.addToFocus': 'Thêm vào tập trung',
+        'projects.removeFromFocus': 'Xóa khỏi tập trung',
+        'projects.maxFocusedProjects': 'Tối đa 5 dự án tập trung',
+        'agenda.inProgress': 'Đang thực hiện',
+        'agenda.overdue': 'Quá hạn',
+        'agenda.dueToday': 'Đến hạn hôm nay',
+        'agenda.starting': 'Đang bắt đầu / Khả dụng',
+        'agenda.nextActions': 'Hành động tiếp theo',
+        'agenda.reviewDue': 'Xem lại đến hạn',
+        'agenda.reviewDueProjects': 'Dự án cần xem lại',
+        'agenda.upcoming': 'Sắp tới',
+        'agenda.allClear': 'Tất cả ổn!',
+        'agenda.noTasks': 'Không có gì trên bàn. Chọn một hành động tiếp theo để tập trung.',
+        'focus.schedule': 'Hôm nay',
+        'focus.nextActions': 'Hành động tiếp theo',
+        'focus.groupBy': 'Nhóm theo',
+        'focus.group.context': 'Ngữ cảnh',
+        'focus.group.project': 'Dự án',
+        'focus.group.area': 'Khu vực',
+        'focus.group.energy': 'Năng lượng',
+        'focus.group.priority': 'Ưu tiên',
+        'focus.group.none': 'Không nhóm',
+        'focus.group.noEnergy': 'Không có năng lượng',
+        'focus.group.noPriority': 'Không có ưu tiên',
+
+        // Task Age
+        'taskAge.day': 'ngày tuổi',
+        'taskAge.days': 'ngày tuổi',
+        'taskAge.week': 'tuần tuổi',
+        'taskAge.weeks': 'tuần tuổi',
+        'taskAge.month': 'tháng tuổi',
+        'taskAge.months': 'tháng tuổi',
+
+        // Energy Contexts
+        'context.energy.focused': 'Tập trung',
+        'context.energy.lowenergy': 'Năng lượng thấp',
+        'context.energy.creative': 'Sáng tạo',
+        'context.energy.routine': 'Thường ngày',
+
+        // Waiting (Mobile)
+        'waiting.title': 'Đang chờ',
+        'waiting.subtitle': 'Nhiệm vụ đang chờ người khác hoặc sự kiện bên ngoài',
+        'waiting.count': 'Đang chờ',
+        'waiting.withDeadline': 'Có hạn chót',
+        'waiting.moveToNext': 'Chuyển sang Tiếp theo',
+        'waiting.markDone': 'Đánh dấu xong',
+        'waiting.empty': 'Không có gì đang chờ',
+        'waiting.emptyHint': 'Đặt nhiệm vụ ở đây khi bước tiếp theo là của người khác — bạn sẽ thấy chúng cho đến khi chúng được giải quyết.',
+
+        // Someday (Mobile)
+        'someday.title': 'Tương lai/Có thể',
+        'someday.subtitle': 'Ý tưởng và mục tiêu bạn có thể muốn theo đuổi trong tương lai',
+        'someday.ideas': 'Ý tưởng',
+        'someday.inProjects': 'Trong dự án',
+        'someday.moveToNext': 'Chuyển sang Tiếp theo',
+        'someday.archive': 'Lưu trữ',
+        'someday.empty': 'Không có gì trong danh sách tương lai',
+        'someday.emptyHint': 'Đặt ý tưởng và mục tiêu bạn có thể theo đuổi sau — ra khỏi bàn, không ra khỏi tâm trí.',
+        'done.emptyHint': 'Các nhiệm vụ đã hoàn thành sẽ xuất hiện ở đây — nhật ký những gì bạn đã hoàn thành.',
+
+        // Search
+        'search.title': 'Tìm kiếm',
+        'search.placeholder': 'Tìm kiếm nhiệm vụ và dự án...',
+        'search.noResults': 'Không tìm thấy kết quả',
+        'search.searching': 'Đang tìm kiếm…',
+        'search.typeToSearch': 'Nhập để tìm kiếm...',
+        'search.resultProject': 'Dự án',
+        'search.resultTask': 'Nhiệm vụ',
+        'search.inProjectSuffix': 'Trong Dự án',
+        'search.showingFirst': 'Đang hiển thị {shown} trong {total} kết quả',
+        'search.saveSearch': 'Lưu tìm kiếm',
+        'search.saveSearchPrompt': 'Đặt tên cho tìm kiếm này',
+        'search.savedSearches': 'Tìm kiếm đã lưu',
+        'search.noSavedSearches': 'Chưa có tìm kiếm đã lưu nào.',
+        'search.deleteConfirm': 'Xóa tìm kiếm đã lưu này?',
+        'search.helpOperators': 'Dùng các toán tử như status:, context:, tag:, project:, due:<=7d.',
+        'search.includeCompleted': 'Bao gồm nhiệm vụ Hoàn thành và Đã lưu trữ',
+        'search.includeReference': 'Bao gồm nhiệm vụ Tham khảo',
+        'search.include.label': 'Bao gồm',
+        'search.scope.label': 'Phạm vi',
+        'search.scope.all': 'Tất cả',
+        'search.scope.projects': 'Chỉ dự án',
+        'search.scope.tasks': 'Chỉ nhiệm vụ',
+        'search.scope.projectTasks': 'Nhiệm vụ trong dự án',
+        'search.due.label': 'Ngày đến hạn',
+        'search.due.any': 'Bất kỳ',
+        'search.due.overdue': 'Quá hạn',
+        'search.due.today': 'Hôm nay',
+        'search.due.tomorrow': 'Ngày mai',
+        'search.due.thisWeek': 'Tuần này',
+        'search.due.nextWeek': 'Tuần tới',
+        'search.due.none': 'Không có ngày đến hạn',
+
+        // Archived
+        'archived.title': 'Đã lưu trữ',
+        'archived.searchPlaceholder': 'Tìm kiếm nhiệm vụ đã lưu trữ...',
+        'archived.noTasksFound': 'Không tìm thấy nhiệm vụ đã lưu trữ nào.',
+        'archived.restoreToInbox': 'Khôi phục về Hộp thư đến',
+        'archived.deletePermanently': 'Xóa vĩnh viễn',
+        'archived.empty': 'Không có nhiệm vụ đã lưu trữ',
+        'archived.emptyHint': 'Các nhiệm vụ bạn lưu trữ sẽ xuất hiện ở đây',
+        'trash.title': 'Thùng rác',
+        'trash.searchPlaceholder': 'Tìm kiếm nhiệm vụ trong thùng rác...',
+        'trash.noTasksFound': 'Không tìm thấy nhiệm vụ trong thùng rác.',
+        'trash.restoreToInbox': 'Khôi phục về Hộp thư đến',
+        'trash.deletePermanently': 'Xóa vĩnh viễn',
+        'trash.empty': 'Thùng rác trống',
+        'trash.emptyHint': 'Các nhiệm vụ đã xóa sẽ xuất hiện ở đây',
+        'trash.deletedAt': 'Đã xóa',
+        'trash.deleteConfirm': 'Xóa vĩnh viễn?',
+        'trash.deleteConfirmBody': 'Hành động này không thể hoàn tác.',
+        'trash.clearAll': 'Xóa thùng rác',
+        'trash.clearAllConfirm': 'Xóa tất cả nhiệm vụ trong thùng rác?',
+        'trash.clearAllConfirmBody': 'Điều này sẽ xóa vĩnh viễn tất cả nhiệm vụ trong thùng rác.',
+        'contexts.search': 'Tìm kiếm ngữ cảnh...',
+        'settings.density': 'Mật độ',
+        'settings.densityDesc': 'Điều chỉnh khoảng cách danh sách để hiển thị nhiều mục hơn trên màn hình.',
+        'settings.densityComfortable': 'Thoải mái',
+        'settings.densityCompact': 'Thu gọn',
+        'settings.globalQuickAddShortcut': 'Phím tắt thêm nhanh toàn cục',
+        'settings.globalQuickAddShortcutDesc': 'Dùng trên toàn hệ thống để ghi lại nhanh, ngay cả khi Mindwtr không được tập trung.',
+        'settings.windowDecorations': 'Trang trí cửa sổ',
+        'settings.windowDecorationsDesc': 'Hiện thanh tiêu đề hệ thống và đường viền cửa sổ.',
+        'settings.showTray': 'Hiện biểu tượng khay hệ thống',
+        'settings.showTrayDesc': 'Hiển thị biểu tượng ứng dụng trong khay hệ thống để truy cập nhanh.',
+        'settings.aiConsentTitle': 'Bật tính năng AI?',
+        'settings.aiConsentDescription': 'Để dùng tính năng này, dữ liệu nhiệm vụ của bạn sẽ được gửi đến {provider} để xử lý. Dữ liệu này tuân theo chính sách quyền riêng tư của họ. Mindwtr không thu thập dữ liệu này. Bạn có muốn tiếp tục không?',
+        'settings.aiConsentCancel': 'Hủy',
+        'settings.aiConsentAgree': 'Đồng ý',
+        'settings.on': 'Bật',
+        'settings.off': 'Tắt',
+        'settings.visible': 'Hiển thị',
+        'settings.hidden': 'Ẩn',
+        'settings.manage': 'Quản lý',
+        'settings.sync': 'Đồng bộ',
+        'settings.syncDescription': 'Chọn cách Mindwtr đồng bộ dữ liệu của bạn. Dropbox là một cú nhấp trong các bản dựng được hỗ trợ; Tệp hoạt động với các thư mục đồng bộ cục bộ; WebDAV và Tự lưu trữ bao gồm lưu trữ tùy chỉnh.',
+        'settings.attachmentsCleanup': 'Dọn dẹp tệp đính kèm',
+        'settings.attachmentsCleanupDesc': 'Xóa các tệp đính kèm đã xóa hoặc mồ côi khỏi thiết bị và bộ nhớ đồng bộ của bạn.',
+        'settings.attachmentsCleanupLastRun': 'Lần dọn dẹp cuối',
+        'settings.attachmentsCleanupNever': 'Chưa bao giờ',
+        'settings.attachmentsCleanupRun': 'Chạy dọn dẹp',
+        'settings.attachmentsCleanupRunning': 'Đang dọn dẹp...',
+        'settings.calendarName': 'Tên',
+        'settings.calendarUrl': 'URL ICS',
+        'settings.calendarAdd': 'Thêm lịch',
+        'settings.calendarRemove': 'Xóa',
+        'settings.calendarSystemTitle': 'Lịch Apple (macOS)',
+        'settings.calendarSystemDesc': 'Đọc sự kiện từ tài khoản Apple Calendar đã đồng bộ bởi macOS (iCloud, Google, Exchange) qua EventKit.',
+        'settings.calendarSystemStatus': 'Quyền truy cập Apple Calendar',
+        'settings.calendarSystemPermissionGranted': 'Đã cấp',
+        'settings.calendarSystemPermissionUndetermined': 'Chưa yêu cầu',
+        'settings.calendarSystemPermissionDenied': 'Bị từ chối',
+        'settings.calendarSystemPermissionUnsupported': 'Không được hỗ trợ',
+        'settings.calendarSystemRequestAccess': 'Yêu cầu quyền truy cập',
+        'settings.calendarSystemDeniedHint': 'Quyền truy cập lịch bị từ chối. Bật Mindwtr trong macOS System Settings > Privacy & Security > Calendars.',
+        'settings.syncFolderLocation': 'Thư mục đồng bộ',
+        'settings.savePath': 'Lưu',
+        'settings.browse': 'Duyệt…',
+        'settings.syncNow': 'Đồng bộ ngay',
+        'settings.syncing': 'Đang đồng bộ…',
+        'settings.syncQueued': 'Đồng bộ đã xếp hàng',
+        'settings.pathHint': 'Chọn thư mục (vd. ~/Sync/mindwtr). Với Syncthing: dùng Gửi & Nhận + Theo dõi thay đổi, sau đó nhấn Đồng bộ ngay trước khi chuyển thiết bị.',
+        'settings.cloudProvider': 'Nhà cung cấp đám mây',
+        'settings.cloudProviderSelfHosted': 'Tự lưu trữ',
+        'settings.cloudProviderDropbox': 'Dropbox',
+        'settings.dropboxAppKey': 'Tài khoản Dropbox',
+        'settings.dropboxAppKeyHint': 'Khóa ứng dụng Dropbox được tiêm vào lúc xây dựng/phát hành.',
+        'settings.dropboxRedirectUri': 'URI chuyển hướng',
+        'settings.dropboxStatus': 'Trạng thái',
+        'settings.dropboxConnected': 'Đã kết nối',
+        'settings.dropboxNotConnected': 'Chưa kết nối',
+        'settings.dropboxConnect': 'Kết nối Dropbox',
+        'settings.dropboxDisconnect': 'Ngắt kết nối Dropbox',
+        'settings.dropboxTest': 'Kiểm tra kết nối',
+        'settings.dropboxTestReachable': 'Có thể truy cập',
+        'settings.dropboxTestFailed': 'Thất bại',
+        'settings.lastSyncConflicts': 'Xung đột',
+        'settings.lastSyncSkew': 'Độ lệch đồng hồ',
+        'settings.lastSyncAdjusted': 'Sửa dấu thời gian',
+        'settings.lastSyncConflictIds': 'ID xung đột',
+        'settings.syncHistoryBackend': 'Nền đồng bộ',
+        'settings.syncHistoryType': 'Loại',
+        'settings.syncHistoryDetails': 'Chi tiết',
+        'settings.syncSettings': 'Cài đặt đồng bộ',
+        'settings.selectFolder': 'Chọn thư mục',
+        'settings.syncStatusFailedSuffix': ' (thất bại)',
+        'settings.syncStatusConflictsSuffix': ' (xung đột)',
+        'settings.fileSyncHowToIos': '1. Đầu tiên, nhấn "Xuất sao lưu" và lưu vào iCloud Drive, hoặc dùng WebDAV để đồng bộ nhiều thiết bị\n2. Nhấn "Chọn thư mục" và chọn thư mục iCloud Drive có thể ghi\n3. Sau đó nhấn "Đồng bộ" để hợp nhất dữ liệu\n\nLưu ý: Các nhà cung cấp Tệp bên thứ ba như Google Drive và OneDrive hiện trả về các bản sao tạm thời trong iOS, vì vậy chúng chưa đáng tin cậy để đồng bộ tệp',
+        'settings.fileSyncHowToAndroid': '1. Đầu tiên, nhấn "Xuất sao lưu" và lưu vào thư mục đồng bộ của bạn (vd. Google Drive)\n2. Nhấn "Chọn thư mục" để cấp quyền truy cập vào thư mục đó\n3. Sau đó nhấn "Đồng bộ" để hợp nhất dữ liệu',
+        'settings.fileSyncTip': 'Mẹo: Để chỉnh sửa nhiều thiết bị thường xuyên, WebDAV được khuyến nghị. Nếu dùng Syncthing, bật Gửi & Nhận + Theo dõi thay đổi và nhấn Đồng bộ trước khi chuyển thiết bị.',
+        'settings.syncReadMergeFolder': 'Đọc và hợp nhất thư mục đồng bộ',
+        'settings.syncReadMergeWebdav': 'Đọc và hợp nhất dữ liệu WebDAV',
+        'settings.syncReadMergeSelfHosted': 'Đọc và hợp nhất dữ liệu tự lưu trữ',
+        'settings.backup': 'Sao lưu',
+        'settings.exportBackup': 'Xuất sao lưu',
+        'settings.saveToSyncFolder': 'Lưu vào thư mục đồng bộ',
+        'settings.syncHistory': 'Lịch sử đồng bộ',
+        'settings.recoverySnapshots': 'Ảnh chụp khôi phục',
+        'settings.recoverySnapshotsDesc': 'Được tạo trước khi đồng bộ. Giữ đến 7 ngày (tối đa 5 tệp).',
+        'settings.recoverySnapshotsLoading': 'Đang tải ảnh chụp…',
+        'settings.recoverySnapshotsEmpty': 'Chưa có ảnh chụp nào.',
+        'settings.recoverySnapshotsRestore': 'Khôi phục',
+        'settings.recoverySnapshotsConfirm': 'Khôi phục ảnh chụp {snapshot}? Điều này sẽ thay thế dữ liệu cục bộ hiện tại.',
+        'settings.recoverySnapshotsConfirmTitle': 'Khôi phục ảnh chụp?',
+        'settings.recoverySnapshotsConfirmCancel': 'Hủy',
+        'settings.developer': 'Nhà phát triển',
+        'settings.website': 'Trang web nhà phát triển',
+        'settings.github': 'GitHub',
+        'settings.documentation': 'Tài liệu',
+        'settings.sponsorProject': 'Tài trợ dự án',
+        'settings.license': 'Giấy phép',
+        'settings.feedback': 'Gửi phản hồi',
+        'settings.feedbackDesc': 'Báo cáo lỗi hoặc gợi ý tính năng. Không cần tài khoản.',
+        'settings.feedbackCategory': 'Danh mục',
+        'settings.feedbackCategoryBug': 'Báo cáo lỗi',
+        'settings.feedbackCategoryFeature': 'Yêu cầu tính năng',
+        'settings.feedbackCategoryOther': 'Khác',
+        'settings.feedbackMessage': 'Tin nhắn',
+        'settings.feedbackMessagePlaceholder': 'Cho chúng tôi biết điều gì đã xảy ra hoặc điều gì sẽ giúp ích.',
+        'settings.feedbackMessagePlaceholderBug': 'Bạn mong đợi gì và điều gì đã xảy ra thay thế? Các bước để tái hiện nếu bạn có thể.',
+        'settings.feedbackMessagePlaceholderFeature': 'Bạn đang cố gắng làm gì, điều gì cản trở và điều gì sẽ giúp ích?',
+        'settings.feedbackMessagePlaceholderOther': 'Cho chúng tôi biết bạn đang nghĩ gì.',
+        'settings.feedbackWhere': 'Điều này xảy ra ở đâu?',
+        'settings.feedbackWherePlaceholder': 'Chọn màn hình hoặc quy trình làm việc (tùy chọn)',
+        'settings.feedbackWhereMessagePrefix': 'Ở đâu',
+        'settings.feedbackWhereInbox': 'Hộp thư đến',
+        'settings.feedbackWhereFocus': 'Tập trung',
+        'settings.feedbackWhereProjects': 'Dự án',
+        'settings.feedbackWhereReview': 'Xem lại',
+        'settings.feedbackWhereSettings': 'Cài đặt',
+        'settings.feedbackWhereSync': 'Đồng bộ',
+        'settings.feedbackWhereImportExport': 'Nhập hoặc xuất',
+        'settings.feedbackWhereNotifications': 'Thông báo',
+        'settings.feedbackWhereOther': 'Khác',
+        'settings.feedbackEmail': 'Email trả lời (tùy chọn)',
+        'settings.feedbackEmailPlaceholder': 'ban@example.com',
+        'settings.feedbackIncludeDiagnostics': 'Bao gồm chẩn đoán gần đây',
+        'settings.feedbackIncludeDiagnosticsDesc': 'Thêm nhật ký ứng dụng đã làm sạch gần đây. Dữ liệu nhiệm vụ, sao lưu và ảnh chụp màn hình không được đính kèm.',
+        'settings.feedbackPrivacy': 'Điều này gửi tin nhắn, phiên bản ứng dụng, nền tảng và chẩn đoán của bạn chỉ nếu bạn chọn bao gồm chúng. Email trả lời chỉ được gửi nếu bạn nhập. Nội dung nhiệm vụ không được đính kèm.',
+        'settings.feedbackSubmit': 'Gửi phản hồi',
+        'settings.feedbackSending': 'Đang gửi...',
+        'settings.feedbackSent': 'Đã gửi phản hồi. Cảm ơn bạn.',
+        'settings.feedbackFailed': 'Không thể gửi phản hồi. Tin nhắn của bạn vẫn ở đây; vui lòng thử lại.',
+        'settings.feedbackUnavailable': 'Phản hồi không được cấu hình trong bản dựng này.',
+        'settings.feedbackUnavailableDesc': 'Bạn vẫn có thể báo cáo lỗi và yêu cầu tính năng với các mẫu vấn đề GitHub.',
+        'settings.feedbackOpenGitHubIssue': 'Mở vấn đề GitHub',
+        'settings.feedbackRequired': 'Thêm tin nhắn ngắn trước.',
+        'settings.feedbackInvalidEmail': 'Nhập email hợp lệ hoặc để trống.',
+        'settings.checkForUpdates': 'Kiểm tra cập nhật',
+        'settings.checkStoreUpdates': 'Kiểm tra Microsoft Store',
+        'settings.checking': 'Đang kiểm tra…',
+        'settings.upToDate': 'Bạn đang dùng phiên bản mới nhất!',
+        'settings.updateAvailable': 'Có cập nhật mới',
+        'settings.checkFailed': 'Kiểm tra cập nhật thất bại',
+        'settings.storeUpdateHint': 'Cập nhật được quản lý bởi Microsoft Store.',
+        'settings.download': 'Tải xuống',
+        'settings.downloadStarting': 'Đang mở tải xuống…',
+        'settings.downloadStarted': 'Tải xuống đã bắt đầu trong trình duyệt của bạn.',
+        'settings.downloadFailed': 'Không thể mở liên kết tải xuống.',
+        'settings.downloadChecksumMismatch': 'Xác minh checksum thất bại. Vui lòng tải xuống lại từ trang phát hành.',
+        'settings.downloadRecommended': 'Gói khuyến nghị',
+        'settings.downloadAURHint': 'Phát hiện Arch: cập nhật qua AUR',
+        'settings.changelog': 'Nhật ký thay đổi',
+        'settings.noChangelog': 'Không có nhật ký thay đổi',
+        'settings.later': 'Sau',
+        'settings.linuxUpdateHint': 'Trên Linux, Mindwtr không thể tự động cài đặt cập nhật. Sau khi tải xuống, cài đặt qua trình quản lý gói của bạn (vd. yay -S mindwtr hoặc paru -S mindwtr) hoặc gói đã tải xuống.',
+        'settings.saved': 'Đã lưu cài đặt',
+        'settings.selectSyncFolderTitle': 'Chọn thư mục đồng bộ',
+        'calendar.mobile.1Day': '1 ngày',
+        'calendar.mobile.day': 'Ngày',
+        'calendar.mobile.dayDetailsPanelHandle': 'Tay cầm bảng chi tiết ngày',
+        'calendar.mobile.end': 'Kết thúc',
+        'calendar.mobile.existingTask': 'Nhiệm vụ hiện có',
+        'calendar.mobile.loading': 'Đang tải…',
+        'calendar.mobile.month': 'Tháng',
+        'calendar.mobile.newTask': 'Nhiệm vụ mới',
+        'calendar.mobile.noMatchingTasks': 'Không có nhiệm vụ nào khớp',
+        'calendar.mobile.scheduleTask': 'Lên lịch nhiệm vụ',
+        'calendar.mobile.show1VisibleDay': 'Hiển thị 1 ngày hiển thị',
+        'calendar.mobile.showFewerDays': 'Hiển thị ít ngày hơn',
+        'calendar.mobile.showMoreDays': 'Hiển thị nhiều ngày hơn',
+        'calendar.mobile.showVisibleDayCount': 'Hiển thị {{dayCount}} ngày hiển thị',
+        'calendar.mobile.swipeUpOrDownToResizeTheDayDetailsPanel': 'Vuốt lên hoặc xuống để thay đổi kích thước bảng chi tiết ngày.',
+        'calendar.mobile.swipeUpOrDownToShowMoreOrFewerDays': 'Vuốt lên hoặc xuống để hiển thị nhiều hoặc ít ngày hơn.',
+        'calendar.mobile.visibleDayCount': '{{dayCount}} ngày',
+        'calendar.mobile.visibleWeekDays': 'Các ngày trong tuần hiển thị',
+        'calendar.mobile.week': 'Tuần',
+        'settings.aboutMobile.couldNotOpenTheAppStoreRatingPagePleaseTry': 'Không thể mở trang đánh giá App Store. Vui lòng thử lại sau.',
+        'settings.aboutMobile.googlePlayCheckWasUnavailableButYourVersionMatchesThe': 'Kiểm tra Google Play không khả dụng, nhưng phiên bản của bạn khớp với phiên bản mới nhất trên GitHub.',
+        'settings.aboutMobile.inAppUpdateChecksAreDisabledInThisFossBuild': 'Kiểm tra cập nhật trong ứng dụng bị tắt trong bản dựng FOSS này. Vui lòng cập nhật từ kho lưu trữ hoặc nguồn gói của bạn.',
+        'settings.aboutMobile.rateOurApp': 'Đánh giá ứng dụng của chúng tôi',
+        'settings.aboutMobile.storeUnavailable': 'Cửa hàng không khả dụng',
+        'settings.aboutMobile.tapToCheck': 'Nhấn để kiểm tra',
+        'settings.aboutMobile.updateIsAvailableOnGooglePlayOpenAppListingNow': 'Có cập nhật trên Google Play. Mở danh sách ứng dụng ngay bây giờ?',
+        'settings.aboutMobile.updatesAreManagedByYourDistributionSource': 'Cập nhật được quản lý bởi nguồn phân phối của bạn',
+        'settings.aboutMobile.upToDate': 'Đã cập nhật',
+        'settings.aboutMobile.googlePlayUpdateAvailableWithVersions': 'v{{currentVersion}} → v{{latestVersion}}\n\nKiểm tra Google Play không khả dụng, nhưng có phiên bản mới hơn trên GitHub. Phát hành Play có thể chậm hơn. Mở danh sách ứng dụng ngay bây giờ?',
+        'settings.aboutMobile.appStoreUpdateAvailableWithVersions': 'v{{currentVersion}} → v{{latestVersion}}\n\nCó cập nhật trên App Store. Mở danh sách ứng dụng ngay bây giờ?',
+        'settings.aboutMobile.youAreUsingTheLatestAppStoreVersion': 'Bạn đang dùng phiên bản App Store mới nhất!',
+        'settings.aboutMobile.youAreUsingTheLatestGooglePlayVersion': 'Bạn đang dùng phiên bản Google Play mới nhất!',
+        'settings.aiMobile.downloadFailedPleaseRetryOnWiFiLargeModelsCannot': 'Tải xuống thất bại. Vui lòng thử lại trên Wi-Fi. Các mô hình lớn không thể được lưu vào bộ nhớ đệm.',
+        'settings.aiMobile.enableAiFeatures': 'Bật tính năng AI?',
+        'settings.aiMobile.localCustomOpenaiCompatible': 'Cục bộ / Tùy chỉnh (tương thích OpenAI)',
+        'settings.aiMobile.localWhisper': 'Whisper cục bộ',
+        'settings.aiMobile.offlineModelPathIsFolder': 'Đường dẫn mô hình ngoại tuyến là thư mục ({{path}}). Vui lòng xóa và thử lại.',
+        'settings.aiMobile.offlineModelPathIsUnsafe': 'Đường dẫn mô hình ngoại tuyến không an toàn để chỉnh sửa ({{path}}).',
+        'settings.aiMobile.suggestions': 'Gợi ý',
+        'settings.aiMobile.toUseAiAssistantYourTaskTextAndOptionalNotes': 'Để dùng trợ lý AI, văn bản nhiệm vụ và ghi chú tùy chọn của bạn sẽ được gửi trực tiếp đến điểm cuối tương thích OpenAI đã cấu hình (ví dụ: máy chủ LLM cục bộ hoặc tự lưu trữ) bằng khóa API của bạn. Mindwtr không thu thập dữ liệu này. Bạn có muốn tiếp tục không?',
+        'settings.aiMobile.aiAssistantPrivacyPromptForProvider': 'Để dùng trợ lý AI, văn bản nhiệm vụ và ghi chú tùy chọn của bạn sẽ được gửi trực tiếp đến {{provider}} bằng khóa API của bạn. Mindwtr không thu thập dữ liệu này. Chính sách quyền riêng tư của nhà cung cấp: {{privacyUrl}}. Bạn có muốn tiếp tục không?',
+        'settings.aiMobile.useTheApiKeyForYourLocalOrSelfHosted': 'Dùng khóa API cho máy chủ tương thích OpenAI cục bộ hoặc tự lưu trữ của bạn nếu cần.',
+        'settings.aiMobile.taskTextSentToProvider': 'Khi được bật, văn bản nhiệm vụ được gửi trực tiếp đến {{provider}} bằng khóa API của bạn.',
+        'settings.aiMobile.whisperDownloadsRequireADevBuildOrProductionBuildNot': 'Tải xuống Whisper yêu cầu bản dựng dev hoặc bản dựng sản xuất (không phải Expo Go).',
+        'settings.aiMobile.whisperTranscriptionRequiresADevBuildOrProductionBuildNot': 'Phiên âm Whisper yêu cầu bản dựng dev hoặc bản dựng sản xuất (không phải Expo Go).',
+        'settings.backupMobile.backupDateLabel': 'Ngày sao lưu: {{backupDate}}',
+        'settings.backupMobile.backupRestoredWithSnapshot': 'Sao lưu khôi phục thành công. Ảnh chụp khôi phục đã lưu là {{snapshotName}}.',
+        'settings.backupMobile.backupPreviewCounts': 'Chứa {{taskCount}} nhiệm vụ và {{projectCount}} dự án.',
+        'settings.backupMobile.failedToExportBackup': 'Xuất sao lưu thất bại',
+        'settings.backupMobile.fileLabel': 'Tệp: {{fileName}}',
+        'settings.backupMobile.import': 'Nhập',
+        'settings.backupMobile.importComplete': 'Nhập hoàn tất',
+        'settings.backupMobile.importDgtGtdData': 'Nhập dữ liệu DGT GTD?',
+        'settings.backupMobile.importTicktickData': 'Nhập dữ liệu TickTick?',
+        'settings.backupMobile.importedTasksKeepOmnifocusNotesDatesTagsRecurrenceAndChecklist': 'Các nhiệm vụ đã nhập giữ ghi chú, ngày, thẻ, lặp lại và danh sách kiểm tra OmniFocus khi được hỗ trợ.',
+        'settings.backupMobile.importedTasksStayInInboxSoYouCanProcessThem': 'Các nhiệm vụ đã nhập ở lại Hộp thư đến để bạn có thể xử lý chúng trong Mindwtr.',
+        'settings.backupMobile.importedTaskProjectCounts': 'Đã nhập {{taskCount}} nhiệm vụ và {{projectCount}} dự án.',
+        'settings.backupMobile.importedTodoistTasksIntoProjects': 'Đã nhập {{taskCount}} nhiệm vụ vào {{projectCount}} dự án.',
+        'settings.backupMobile.importedTaskProjectAreaCounts': 'Đã nhập {{taskCount}} nhiệm vụ, {{projectCount}} dự án và {{areaCount}} khu vực.',
+        'settings.backupMobile.importFailed': 'Nhập thất bại',
+        'settings.backupMobile.importOmnifocusData': 'Nhập dữ liệu OmniFocus?',
+        'settings.backupMobile.importTodoistData': 'Nhập dữ liệu Todoist?',
+        'settings.backupMobile.importTasksFromFile': 'Nhập {{taskCount}} nhiệm vụ từ {{fileName}}?',
+        'settings.backupMobile.importTaskCountFromFile': 'Nhập {{taskCount}} nhiệm vụ từ {{fileName}}?',
+        'settings.backupMobile.importTodoistTasksFromProjects': 'Nhập {{taskCount}} nhiệm vụ từ {{projectCount}} dự án Todoist?',
+        'settings.backupMobile.invalidBackup': 'Sao lưu không hợp lệ',
+        'settings.backupMobile.recoverySnapshotRestored': 'Đã khôi phục ảnh chụp khôi phục.',
+        'settings.backupMobile.recoverySnapshotSaved': 'Đã lưu ảnh chụp khôi phục là {{snapshotName}}.',
+                'settings.backupMobile.restoreBackup': 'Khôi phục sao lưu?',
+        'settings.backupMobile.restoreComplete': 'Khôi phục hoàn tất',
+        'settings.backupMobile.restoreFailed': 'Khôi phục thất bại',
+        'settings.backupMobile.restoreRecoverySnapshot': 'Khôi phục ảnh chụp khôi phục?',
+        'settings.backupMobile.restoreSnapshotReplaceLocalData': 'Khôi phục {{snapshotName}}? Điều này sẽ thay thế dữ liệu cục bộ hiện tại.',
+        'settings.backupMobile.theSelectedFileIsNotASupportedDgtGtdExport': 'Tệp đã chọn không phải là xuất DGT GTD được hỗ trợ.',
+        'settings.backupMobile.theSelectedFileIsNotASupportedOmnifocusExport': 'Tệp đã chọn không phải là xuất OmniFocus được hỗ trợ.',
+        'settings.backupMobile.theSelectedFileIsNotASupportedTodoistExport': 'Tệp đã chọn không phải là xuất Todoist được hỗ trợ.',
+        'settings.backupMobile.theSelectedFileIsNotASupportedTicktickBackup': 'Tệp đã chọn không phải là sao lưu TickTick được hỗ trợ.',
+        'settings.backupMobile.thisFileIsNotAValidMindwtrBackup': 'Tệp này không phải là sao lưu Mindwtr hợp lệ.',
+        'settings.backupMobile.thisWillReplaceAllCurrentLocalDataARecoverySnapshot': 'Điều này sẽ thay thế tất cả dữ liệu cục bộ hiện tại. Ảnh chụp khôi phục sẽ được lưu trước.',
+        'settings.backupMobile.omnifocusAreasCreated': '{{areaCount}} khu vực đã được tạo từ các thư mục OmniFocus.',
+        'settings.backupMobile.dgtAreasWillBeCreated': '{{areaCount}} khu vực sẽ được tạo từ các thư mục DGT.',
+        'settings.backupMobile.ticktickAreasWillBeCreated': '{{areaCount}} khu vực sẽ được tạo từ các thư mục TickTick.',
+        'settings.backupMobile.ticktickProjectsWillBeCreated': '{{projectCount}} dự án sẽ được tạo từ các danh sách TickTick.',
+        'settings.backupMobile.omnifocusAreasWillBeCreated': '{{areaCount}} khu vực sẽ được tạo từ các thư mục OmniFocus khi cần.',
+        'settings.backupMobile.checklistItemsPreserved': '{{checklistItemCount}} mục danh sách kiểm tra đã được giữ lại.',
+        'settings.backupMobile.checklistItemsWillBePreserved': '{{checklistItemCount}} mục danh sách kiểm tra sẽ được giữ lại.',
+        'settings.backupMobile.moreProjects': '• {{projectCount}} dự án nữa…',
+        'settings.backupMobile.nestedTasksBecameChecklistItems': '{{taskCount}} nhiệm vụ lồng nhau đã trở thành mục danh sách kiểm tra.',
+        'settings.backupMobile.nestedTasksWillBecomeChecklistItems': '{{taskCount}} nhiệm vụ lồng nhau sẽ trở thành mục danh sách kiểm tra khi có thể.',
+        'settings.backupMobile.projectsWillBeCreated': '{{projectCount}} dự án sẽ được tạo.',
+        'settings.backupMobile.projectsWillBeCreatedWhenNeeded': '{{projectCount}} dự án sẽ được tạo khi cần.',
+        'settings.backupMobile.recurringTasksWillKeepSupportedRepeatRules': '{{taskCount}} nhiệm vụ lặp lại sẽ giữ các quy tắc lặp lại được hỗ trợ.',
+        'settings.backupMobile.sectionsWillBePreserved': '{{sectionCount}} phần sẽ được giữ lại.',
+        'settings.backupMobile.subtasksBecameChecklistItems': '{{subtaskCount}} nhiệm vụ con đã trở thành mục danh sách kiểm tra.',
+        'settings.backupMobile.subtasksWillBecomeChecklistItems': '{{subtaskCount}} nhiệm vụ con sẽ trở thành mục danh sách kiểm tra.',
+        'settings.backupMobile.tasksStayedOutsideProjects': '{{taskCount}} nhiệm vụ vẫn ở ngoài dự án.',
+        'settings.backupMobile.tasksWillStayOutsideProjects': '{{taskCount}} nhiệm vụ sẽ ở ngoài dự án để bạn có thể xử lý chúng trong Mindwtr.',
+
+        // Apple Reminders Import
+        'settings.appleRemindersImport.appleReminders': 'Apple Nhắc nhở',
+        'settings.appleRemindersImport.captureList': 'Danh sách ghi lại',
+        'settings.appleRemindersImport.chooseCaptureList': 'Chọn danh sách Nhắc nhở',
+        'settings.appleRemindersImport.deleteAfterImport': 'Xóa nhắc nhở đã nhập',
+        'settings.appleRemindersImport.deleteAfterImportDescription': 'Xóa nhắc nhở khỏi Apple Nhắc nhở sau khi Mindwtr thêm chúng vào Hộp thư đến.',
+        'settings.appleRemindersImport.deleteFailedCount': '{{taskCount}} nhắc nhở đã nhập không thể xóa khỏi Apple Nhắc nhở.',
+        'settings.appleRemindersImport.deletedCount': 'Đã xóa {{taskCount}} nhắc nhở đã nhập khỏi Apple Nhắc nhở.',
+        'settings.appleRemindersImport.failedCount': '{{taskCount}} nhắc nhở nhập thất bại.',
+        'settings.appleRemindersImport.importedCount': 'Đã nhập {{taskCount}} nhắc nhở vào Hộp thư đến.',
+        'settings.appleRemindersImport.importIncomplete': 'Nhập nhắc nhở chưa hoàn thành',
+        'settings.appleRemindersImport.importIncompleteDeleteDescription': 'Thêm nhắc nhở chưa hoàn thành mới vào Hộp thư đến, sau đó xóa nhắc nhở đã nhập khỏi Apple Nhắc nhở.',
+        'settings.appleRemindersImport.importIncompleteDescription': 'Thêm nhắc nhở chưa hoàn thành mới vào Hộp thư đến. Nhắc nhở vẫn ở trong Apple Nhắc nhở.',
+        'settings.appleRemindersImport.noListsFound': 'Không tìm thấy danh sách Apple Nhắc nhở nào trên thiết bị này.',
+        'settings.appleRemindersImport.nothingNew': 'Không có gì mới để nhập',
+        'settings.appleRemindersImport.permissionRequired': 'Cần quyền truy cập Nhắc nhở để nhập Apple Nhắc nhở.',
+        'settings.appleRemindersImport.selected': 'Đã chọn',
+        'settings.appleRemindersImport.skippedCount': 'Đã bỏ qua {{taskCount}} nhắc nhở đã nhập, đã hoàn thành hoặc thiếu tiêu đề.',
+
+        // Calendar Mobile Settings
+        'settings.calendarMobile.calendarAccessIsRequiredToPushTasksToYourCalendar': 'Cần quyền truy cập lịch để đẩy nhiệm vụ lên lịch của bạn.',
+        'settings.calendarMobile.calendarAccessWasDeniedPleaseGrantAccessInSettings': 'Quyền truy cập lịch bị từ chối. Vui lòng cấp quyền truy cập trong Cài đặt.',
+        'settings.calendarMobile.calendarDeleted': 'Đã xóa lịch',
+        'settings.calendarMobile.calendarSyncDisabled': 'Đã tắt đồng bộ lịch',
+        'settings.calendarMobile.calendarTargetUpdated': 'Đã cập nhật đích lịch',
+        'settings.calendarMobile.chooseAnAccountCalendarIfYourCalendarAppHidesLocal': 'Chọn lịch tài khoản nếu ứng dụng lịch của bạn ẩn lịch cục bộ.',
+        'settings.calendarMobile.dedicatedAccountCalendar': 'Lịch tài khoản chuyên dụng',
+        'settings.calendarMobile.dedicatedLocalCalendar': 'Lịch cục bộ chuyên dụng',
+        'settings.calendarMobile.deleteMindwtrCalendar': 'Xóa lịch Mindwtr',
+        'settings.calendarMobile.dueDateTasksWillBeWrittenToTheSelectedCalendar': 'Các nhiệm vụ có ngày đến hạn sẽ được ghi vào lịch đã chọn.',
+        'settings.calendarMobile.failedToLoadDeviceCalendarSettings': 'Tải cài đặt lịch thiết bị thất bại',
+        'settings.calendarMobile.failedToLoadEvents': 'Tải sự kiện thất bại',
+        'settings.calendarMobile.failedToLoadSavedCalendars': 'Tải lịch đã lưu thất bại',
+        'settings.calendarMobile.failedToLoadWritableCalendars': 'Tải lịch có thể ghi thất bại',
+        'settings.calendarMobile.forASeparateColorInGoogleCalendarSelectADedicated': 'Để có màu riêng trong Google Calendar, chọn lịch tài khoản chuyên dụng tên Mindwtr. Lịch dùng chung vẫn dùng tiền tố tiêu đề Mindwtr.',
+        'settings.calendarMobile.icsSubscriptions': 'Đăng ký ICS',
+        'settings.calendarMobile.chooseLocalIcsFile': 'Chọn tệp .ics cục bộ',
+        'settings.calendarMobile.localIcsFileAdded': 'Đã thêm tệp .ics cục bộ',
+        'settings.calendarMobile.localIcsFilesAreReadOnly': 'Lịch .ics cục bộ chỉ đọc trong Mindwtr.',
+        'settings.calendarMobile.localCalendarTargetsStayOnThisDeviceUseAGoogle': 'Đích lịch cục bộ ở lại trên thiết bị này. Dùng đích lịch tài khoản Google nếu bạn cần đồng bộ web Google Calendar.',
+        'settings.calendarMobile.mindwtrCalendar': 'Lịch Mindwtr',
+        'settings.calendarMobile.optional': 'Tùy chọn',
+        'settings.calendarMobile.permissionRequired': 'Cần quyền truy cập',
+        'settings.calendarMobile.pushTasksToCalendar': 'Đẩy nhiệm vụ lên lịch',
+        'settings.calendarMobile.refreshCalendars': 'Làm mới lịch',
+        'settings.calendarMobile.reloadTheListAfterAddingACalendarInGoogleCalendar': 'Tải lại danh sách sau khi thêm lịch trong Google Calendar.',
+        'settings.calendarMobile.removeTheDedicatedCalendarAndItsPushedEventsFromThis': 'Xóa lịch chuyên dụng và các sự kiện đã đẩy khỏi thiết bị này.',
+        'settings.calendarMobile.scheduledTasksAndTasksWithDueDatesAreAddedTo': 'Các nhiệm vụ đã lên lịch và nhiệm vụ có ngày đến hạn được thêm vào lịch thiết bị đã chọn của bạn.',
+        'settings.calendarMobile.sharedAccountCalendar': 'Lịch tài khoản dùng chung',
+        'settings.calendarMobile.sharedLocalCalendar': 'Lịch cục bộ dùng chung',
+        'settings.calendarMobile.syncTarget': 'Đích đồng bộ',
+        'settings.calendarMobile.tasksWillNoLongerBePushedToYourCalendarExisting': 'Nhiệm vụ sẽ không còn được đẩy lên lịch của bạn. Các sự kiện hiện có được giữ lại.',
+        'settings.calendarMobile.test': 'Kiểm tra',
+        'settings.calendarMobile.theMindwtrCalendarAndAllItsEventsHaveBeenRemoved': 'Lịch Mindwtr và tất cả sự kiện của nó đã bị xóa.',
+
+        // GTD Mobile Settings
+        'settings.gtdMobile.autoStartBreaks': 'Tự động bắt đầu nghỉ',
+        'settings.gtdMobile.autoStartFocus': 'Tự động bắt đầu tập trung',
+        'settings.gtdMobile.captureDefaults': 'Mặc định ghi lại',
+        'settings.gtdMobile.chooseAPresetThenOpenASectionToFineTune': 'Chọn cài đặt sẵn, sau đó mở phần để tinh chỉnh các trường.',
+        'settings.gtdMobile.chooseWhichOptionalStepsAppearInDailyAndWeeklyReview': 'Chọn các bước tùy chọn nào xuất hiện trong xem lại hàng ngày và hàng tuần.',
+        'settings.gtdMobile.currentLayoutCustom': 'Bố cục hiện tại: Tùy chỉnh',
+        'settings.gtdMobile.customPresetTaskLinkingAndAutoStartBehavior': 'Cài đặt sẵn tùy chỉnh, liên kết nhiệm vụ và hành vi tự động bắt đầu.',
+        'settings.gtdMobile.defaultScheduleTime': 'Thời gian lên lịch mặc định',
+        'settings.gtdMobile.full': 'Đầy đủ',
+        'settings.gtdMobile.hhMm': 'GG:PP',
+        'settings.gtdMobile.hideFromEditor': 'Ẩn khỏi trình chỉnh sửa',
+        'settings.gtdMobile.moveToSection': 'Chuyển đến phần',
+        'settings.gtdMobile.optionalPreFillsManualStartDueAndReviewTimeFields': 'Tùy chọn. Điền trước các trường thời gian Bắt đầu, Đến hạn và Xem lại thủ công sau khi bạn chọn ngày. Để trống cho chỉ ngày.',
+        'settings.gtdMobile.orderWithinSection': 'Thứ tự trong phần',
+        'settings.gtdMobile.pomodoroSettings': 'Cài đặt Pomodoro',
+        'settings.gtdMobile.pomodoroWillNowAdvancePhasesAutomatically': 'Pomodoro bây giờ sẽ tự động chuyển sang các giai đoạn tiếp theo.',
+        'settings.gtdMobile.presets': 'Cài đặt sẵn',
+        'settings.gtdMobile.reviewSteps': 'Các bước xem lại',
+        'settings.gtdMobile.showInEditor': 'Hiện trong trình chỉnh sửa',
+        'settings.gtdMobile.simple': 'Đơn giản',
+        'settings.gtdMobile.standard': 'Tiêu chuẩn',
+        'settings.gtdMobile.startTaskEditingWithThisSectionExpanded': 'Bắt đầu chỉnh sửa nhiệm vụ với phần này được mở rộng.',
+        'settings.gtdMobile.startTheBreakTimerAutomaticallyWhenAFocusSessionEnds': 'Tự động bắt đầu hẹn giờ nghỉ khi phiên tập trung kết thúc.',
+        'settings.gtdMobile.startTheNextFocusSessionAutomaticallyWhenABreakEnds': 'Tự động bắt đầu phiên tập trung tiếp theo khi giờ nghỉ kết thúc.',
+        'settings.gtdMobile.useHhMmForTheDefaultScheduleTime': 'Dùng GG:PP cho thời gian lên lịch mặc định.',
+        'settings.mobile.displayHowLongAgoATaskWasCreatedInTask': 'Hiển thị thời gian tạo nhiệm vụ trong siêu dữ liệu nhiệm vụ.',
+        'settings.mobile.quickAccessView': 'Chế độ xem truy cập nhanh',
+        'settings.mobile.showTaskAge': 'Hiện tuổi nhiệm vụ',
+
+        // Sync Mobile Settings
+        'settings.syncMobile.accountStatus': 'Trạng thái tài khoản',
+        'settings.syncMobile.addThisExactRedirectUriInDropboxOauthSettings': 'Thêm URI chuyển hướng chính xác này vào cài đặt Dropbox OAuth.',
+        'settings.syncMobile.backgroundSync': 'Đồng bộ nền',
+        'settings.syncMobile.clearPendingAttachmentDeletes': 'Xóa các lần xóa tệp đính kèm đang chờ?',
+        'settings.syncMobile.cloudkitIsRestrictedOnThisDeviceCheckScreenTimeMdm': 'CloudKit bị hạn chế trên thiết bị này. Kiểm tra Screen Time, MDM hoặc hạn chế iCloud, sau đó thử lại.',
+        'settings.syncMobile.connectedToDropbox': 'Đã kết nối với Dropbox.',
+        'settings.syncMobile.connectionFailed': 'Kết nối thất bại',
+        'settings.syncMobile.connectionOk': 'Kết nối OK',
+        'settings.syncMobile.disconnected': 'Đã ngắt kết nối',
+        'settings.syncMobile.disconnectFailed': 'Ngắt kết nối thất bại',
+        'settings.syncMobile.dropboxAccountIsReachable': 'Tài khoản Dropbox có thể truy cập.',
+        'settings.syncMobile.dropboxAppKeyIsNotConfiguredForThisBuild': 'Khóa ứng dụng Dropbox chưa được cấu hình cho bản dựng này.',
+        'settings.syncMobile.dropboxAppKeyIsNotConfiguredInThisBuild': 'Khóa ứng dụng Dropbox chưa được cấu hình trong bản dựng này.',
+        'settings.syncMobile.dropboxConnectionRemoved': 'Đã xóa kết nối Dropbox.',
+        'settings.syncMobile.dropboxIsDisabledInFossBuilds': 'Dropbox bị tắt trong các bản dựng FOSS.',
+        'settings.syncMobile.dropboxOauthRequiresADevelopmentReleaseBuildExpoGoUses': 'Dropbox OAuth yêu cầu bản dựng phát triển/phát hành. Expo Go dùng các URI chuyển hướng tạm thời mà Dropbox từ chối.',
+        'settings.syncMobile.dropboxTokenIsInvalidOrRevokedPleaseTapConnectDropbox': 'Token Dropbox không hợp lệ hoặc đã bị thu hồi. Vui lòng nhấn Kết nối Dropbox để xác thực lại.',
+        'settings.syncMobile.dropboxUnavailable': 'Dropbox không khả dụng',
+        'settings.syncMobile.dropboxUnavailableInExpoGo': 'Dropbox không khả dụng trong Expo Go',
+        'settings.syncMobile.error': 'Lỗi',
+        'settings.syncMobile.expoGoIsNotSupportedForDropboxOauthUseA': 'Expo Go không được hỗ trợ cho Dropbox OAuth. Dùng bản dựng phát triển/phát hành.',
+        'settings.syncMobile.failedToSaveSelfHostedSettings': 'Lưu cài đặt tự lưu trữ thất bại',
+        'settings.syncMobile.failedToSaveWebdavSettings': 'Lưu cài đặt WebDAV thất bại',
+        'settings.syncMobile.failedToSetSyncPath': 'Đặt đường dẫn đồng bộ thất bại',
+        'settings.syncMobile.howToSync': 'Cách đồng bộ',
+        'settings.syncMobile.httpsRequired': 'Yêu cầu HTTPS',
+        'settings.syncMobile.icloudIsTemporarilyUnavailableWaitAMomentThenTapSync': 'iCloud tạm thời không khả dụng. Chờ một lúc, sau đó nhấn "Đồng bộ ngay" lần nữa.',
+        'settings.syncMobile.icloudRestricted': 'iCloud bị hạn chế',
+        'settings.syncMobile.icloudSignInRequired': 'Cần đăng nhập iCloud',
+        'settings.syncMobile.icloudStatusUnavailable': 'Trạng thái iCloud không khả dụng',
+        'settings.syncMobile.icloudTemporarilyUnavailable': 'iCloud tạm thời không khả dụng',
+        'settings.syncMobile.icloudUnavailable': 'iCloud không khả dụng',
+        'settings.syncMobile.importDgtGtdJsonOrZipExportsIntoMindwtrAreas': 'Nhập các xuất JSON hoặc ZIP DGT GTD vào các khu vực, dự án và nhiệm vụ Mindwtr.',
+        'settings.syncMobile.importFromDgtGtd': 'Nhập từ DGT GTD',
+        'settings.syncMobile.importFromOmnifocus': 'Nhập từ OmniFocus',
+        'settings.syncMobile.importFromTodoist': 'Nhập từ Todoist',
+        'settings.syncMobile.importFromTicktick': 'Nhập từ TickTick',
+        'settings.syncMobile.importOmnifocusCsvJsonOrZipExportsIntoMindwtrProjects': 'Nhập các xuất CSV, JSON hoặc ZIP OmniFocus vào các dự án Mindwtr và nhiệm vụ hộp thư đến.',
+        'settings.syncMobile.importTodoistCsvOrZipExportsIntoMindwtrProjects': 'Nhập các xuất CSV hoặc ZIP Todoist vào các dự án Mindwtr.',
+        'settings.syncMobile.importTicktickCsvOrZipBackupsIntoMindwtrAreas': 'Nhập các sao lưu CSV hoặc ZIP TickTick vào các khu vực, dự án và nhiệm vụ Mindwtr.',
+        'settings.syncMobile.insecureHttpEnabled': 'HTTP không an toàn đã bật',
+        'settings.syncMobile.invalidRedirectUri': 'URI chuyển hướng không hợp lệ',
+        'settings.syncMobile.invalidSyncFile': 'Tệp đồng bộ không hợp lệ',
+        'settings.syncMobile.invalidUrl': 'URL không hợp lệ',
+        'settings.syncMobile.mindwtrAsksTheSystemToSyncAboutEvery15Minutes': 'Mindwtr yêu cầu hệ thống đồng bộ khoảng 15 phút một lần khi điều kiện mạng và pin cho phép. iOS có thể đợi lâu hơn; tối ưu hóa pin Android có thể trì hoãn công việc nền.',
+        'settings.syncMobile.oauthWithDropboxAppFolderAccessMindwtrSyncsAppsMindwtr': 'OAuth với quyền truy cập Thư mục ứng dụng Dropbox. Mindwtr đồng bộ /Apps/Mindwtr/data.json và /Apps/Mindwtr/attachments/* trong Dropbox của bạn.',
+        'settings.syncMobile.onlyClearTheseIfYouNoLongerWantMindwtrTo': 'Chỉ xóa những thứ này nếu bạn không còn muốn Mindwtr xóa các tệp đó khỏi backend đồng bộ cũ. Điều này không xóa bất kỳ tệp nào.',
+        'settings.syncMobile.onlyUseThisOnTrustedNetworksSyncDataWillBe': 'Chỉ dùng trên các mạng đáng tin cậy. Dữ liệu đồng bộ sẽ được gửi không mã hóa.',
+        'settings.syncMobile.openDropboxOauthSignInInBrowser': 'Mở đăng nhập Dropbox OAuth trong trình duyệt.',
+        'settings.syncMobile.pendingRemoteDeletes': 'Xóa từ xa đang chờ',
+        'settings.syncMobile.pleaseChooseAMindwtrBackupJsonFileInTheTarget': 'Vui lòng chọn tệp JSON sao lưu Mindwtr trong thư mục đích, sau đó thử "Chọn thư mục" lại.',
+        'settings.syncMobile.pleaseConnectDropboxFirst': 'Vui lòng kết nối Dropbox trước.',
+        'settings.syncMobile.pleaseEnterAValidSelfHostedUrlHttpHttps': 'Vui lòng nhập URL tự lưu trữ hợp lệ (http/https).',
+        'settings.syncMobile.pleaseEnterAValidWebdavUrlHttpHttps': 'Vui lòng nhập URL WebDAV hợp lệ (http/https).',
+        'settings.syncMobile.pleaseSetASelfHostedUrlFirst': 'Vui lòng đặt URL tự lưu trữ trước',
+        'settings.syncMobile.pleaseSetASyncFolderFirst': 'Vui lòng đặt thư mục đồng bộ trước',
+        'settings.syncMobile.pleaseSetAWebdavUrlFirst': 'Vui lòng đặt URL WebDAV trước',
+        'settings.syncMobile.publicHttpSyncUrlsAreBlockedUseHttpsOrEnable': 'URL đồng bộ HTTP công khai bị chặn. Dùng HTTPS, hoặc bật HTTP không an toàn chỉ cho mạng riêng đáng tin cậy.',
+        'settings.syncMobile.readAndMergeDropboxData': 'Đọc và hợp nhất dữ liệu Dropbox.',
+        'settings.syncMobile.readAndMergeTheLatestCloudkitDataNow': 'Đọc và hợp nhất dữ liệu CloudKit mới nhất ngay bây giờ.',
+        'settings.syncMobile.replaceLocalDataFromABackupJsonFile': 'Thay thế dữ liệu cục bộ từ tệp JSON sao lưu.',
+        'settings.syncMobile.requiresDevelopmentReleaseBuildExpoGoUnsupported': 'Yêu cầu bản dựng phát triển/phát hành (Expo Go không được hỗ trợ).',
+        'settings.syncMobile.restoreBackup': 'Khôi phục sao lưu',
+        'settings.syncMobile.revokeAppTokenAndRemoveLocalAuth': 'Thu hồi token ứng dụng và xóa xác thực cục bộ.',
+        'settings.syncMobile.savedAutomaticallyBeforeRestoreAndImportOperations': 'Được lưu tự động trước các thao tác khôi phục và nhập.',
+        'settings.syncMobile.scheduledBackgroundSyncIsAvailableForWebdavSelfHostedCloud': 'Đồng bộ nền theo lịch có sẵn cho WebDAV, đám mây tự lưu trữ, Dropbox và iCloud. Đồng bộ tệp vẫn chạy từ nền trước và đồng bộ thủ công.',
+        'settings.syncMobile.selfHostedEndpointIsReachable': 'Điểm cuối tự lưu trữ có thể truy cập.',
+        'settings.syncMobile.signedInToIcloud': 'Đã đăng nhập iCloud',
+        'settings.syncMobile.statusConnected': 'Trạng thái: Đã kết nối',
+        'settings.syncMobile.statusNotConnected': 'Trạng thái: Chưa kết nối',
+        'settings.syncMobile.syncFolderIsReadOnly': 'Thư mục đồng bộ chỉ đọc',
+        'settings.syncMobile.syncFolderSetSuccessfully': 'Đã đặt thư mục đồng bộ thành công',
+        'settings.syncMobile.syncsYourTasksProjectsAndAreasAcrossAppleDevicesUsing': 'Đồng bộ nhiệm vụ, dự án và khu vực của bạn trên các thiết bị Apple bằng CloudKit. Không cần thiết lập tài khoản Mindwtr. Nhấn "Đồng bộ ngay" để buộc hợp nhất ngay lập tức.',
+        'settings.syncMobile.syncsYourTasksProjectsAndAreasAcrossAppleDevicesUsing2': 'Đồng bộ nhiệm vụ, dự án và khu vực của bạn trên các thiết bị Apple bằng CloudKit. Nếu đồng bộ không bắt đầu, xác minh rằng iCloud đã bật cho thiết bị và ứng dụng này, sau đó nhấn "Đồng bộ ngay".',
+        'settings.syncMobile.theSelectedFileCameFromATemporaryIosFilesCopy': 'Tệp đã chọn đến từ bản sao tệp iOS tạm thời. Các nhà cung cấp như Google Drive và OneDrive chưa đáng tin cậy để đồng bộ tệp ở đây. Vui lòng chọn iCloud Drive thay thế, hoặc chuyển sang WebDAV.',
+        'settings.syncMobile.theSelectedFileCameFromATemporaryIosFilesCopy2': 'Tệp đã chọn đến từ bản sao tệp iOS tạm thời. Các nhà cung cấp như Google Drive và OneDrive chưa đáng tin cậy để đồng bộ tệp ở đây. Vui lòng đến Cài đặt → Đồng bộ, chọn iCloud Drive, hoặc chuyển sang WebDAV.',
+        'settings.syncMobile.theSelectedFolderIsReadOnlyChooseAWritableLocation': 'Thư mục đã chọn chỉ đọc. Chọn vị trí có thể ghi, hoặc làm thư mục đám mây khả dụng ngoại tuyến trong Tệp trước khi chọn nó.',
+        'settings.syncMobile.theSelectedFolderIsReadOnlyPleaseChooseAWritable': 'Thư mục đã chọn chỉ đọc. Vui lòng chọn thư mục có thể ghi (vd. Tệp của tôi) hoặc làm khả dụng ngoại tuyến.',
+        'settings.syncMobile.thisDeviceIsNotSignedIntoIcloudOpenIosSettings': 'Thiết bị này chưa đăng nhập iCloud. Mở Cài đặt iOS, đăng nhập vào Tài khoản Apple, bật iCloud cho Mindwtr, sau đó quay lại và nhấn "Đồng bộ ngay".',
+        'settings.syncMobile.unsupportedCloudProviderOnIos': 'Nhà cung cấp đám mây không được hỗ trợ trên iOS',
+        'settings.syncMobile.useRedirectUri': 'Dùng URI chuyển hướng',
+        'settings.syncMobile.webdavEndpointIsReachable': 'Điểm cuối WebDAV có thể truy cập.',
+
+        // Mind Sweep (guided capture)
+        'mindSweep.launchButton': 'Quét tâm trí',
+        'mindSweep.title': 'Quét tâm trí',
+        'mindSweep.intro': 'Đi qua một vài danh sách gợi ý và viết xuống bất cứ điều gì nổi lên. Đừng lọc hay sắp xếp — mọi mục đều đổ vào Hộp thư đến để sắp xếp sau.',
+        'mindSweep.scopeLabel': 'Bạn muốn quét gì?',
+        'mindSweep.scopeAll': 'Tất cả mọi thứ',
+        'mindSweep.scopePersonal': 'Cá nhân',
+        'mindSweep.scopeWork': 'Công việc',
+        'mindSweep.start': 'Bắt đầu quét',
+        'mindSweep.back': 'Quay lại',
+        'mindSweep.next': 'Tiếp theo',
+        'mindSweep.finish': 'Hoàn tất',
+        'mindSweep.close': 'Đóng',
+        'mindSweep.inputPlaceholder': 'Viết bất cứ điều gì nảy ra trong đầu…',
+        'mindSweep.add': 'Thêm',
+        'mindSweep.progress': '{{current}} trong {{total}}',
+        'mindSweep.groupCaptured': 'Đã ghi lại trong danh sách này:',
+        'mindSweep.summaryTitle': 'Quét hoàn tất',
+        'mindSweep.summaryCount': 'Bạn đã ghi lại {{count}} mục vào Hộp thư đến.',
+        'mindSweep.summaryEmpty': 'Lần này không có gì nổi lên — đầu bạn có thể đã rỗng rồi.',
+        'mindSweep.summaryHint': 'Xử lý chúng từ Hộp thư đến khi nào bạn sẵn sàng.',
+        'mindSweep.group.homeStuff.title': 'Nhà & đồ đạc',
+        'mindSweep.group.homeStuff.p1': 'Bất cứ thứ gì ở nhà bị hỏng, rò rỉ hoặc phát ra tiếng lạ',
+        'mindSweep.group.homeStuff.p2': 'Thứ gì đó bạn cứ định thay hoặc nâng cấp',
+        'mindSweep.group.homeStuff.p3': 'Ngăn kéo, tủ hoặc góc nào đó cần dọn dẹp',
+        'mindSweep.group.homeStuff.p4': 'Những thứ bạn đã mượn hoặc cho mượn chưa trả lại',
+        'mindSweep.group.homeStuff.p5': 'Dự án nhà chỉ tồn tại trong đầu bạn cho đến nay',
+        'mindSweep.group.peopleLife.title': 'Gia đình & bạn bè',
+        'mindSweep.group.peopleLife.p1': 'Ai đó bạn nợ một cuộc gọi, tin nhắn hoặc lời cảm ơn',
+        'mindSweep.group.peopleLife.p2': 'Sinh nhật, kỷ niệm hoặc dịp sắp tới',
+        'mindSweep.group.peopleLife.p3': 'Lời hứa bạn đã đưa ra và chưa thực hiện',
+        'mindSweep.group.peopleLife.p4': 'Ai đó bạn cứ định gặp để tâm sự',
+        'mindSweep.group.peopleLife.p5': 'Quà cần tìm hoặc buổi tụ tập cần lên kế hoạch',
+        'mindSweep.group.selfCare.title': 'Sức khỏe & bản thân',
+        'mindSweep.group.selfCare.p1': 'Các cuộc hẹn cần đặt — bác sĩ, nha sĩ, cắt tóc',
+        'mindSweep.group.selfCare.p2': 'Thuốc hoặc đơn thuốc sắp hết',
+        'mindSweep.group.selfCare.p3': 'Thói quen bạn muốn bắt đầu hoặc muốn dừng',
+        'mindSweep.group.selfCare.p4': 'Bất cứ điều gì về giấc ngủ, thức ăn hoặc tập thể dục đang làm phiền bạn',
+        'mindSweep.group.selfCare.p5': 'Điều gì đó đang lặng lẽ lo lắng bạn',
+        'mindSweep.group.lifeAdmin.title': 'Tiền & giấy tờ',
+        'mindSweep.group.lifeAdmin.p1': 'Hóa đơn hoặc thanh toán bạn không chắc đã xử lý',
+        'mindSweep.group.lifeAdmin.p2': 'Đăng ký dịch vụ bạn định hủy',
+        'mindSweep.group.lifeAdmin.p3': 'Bảo hiểm, hợp đồng hoặc gia hạn sắp đến hạn',
+        'mindSweep.group.lifeAdmin.p4': 'Biểu mẫu, thuế hoặc đơn xin đang chờ bạn',
+        'mindSweep.group.lifeAdmin.p5': 'Hoàn tiền, khiếu nại hoặc trả hàng bạn chưa theo đuổi',
+        'mindSweep.group.funGrowth.title': 'Vui chơi & học hỏi',
+        'mindSweep.group.funGrowth.p1': 'Chuyến đi bạn cứ nói sẽ lên kế hoạch',
+        'mindSweep.group.funGrowth.p2': 'Sách, phim hoặc nhạc người khác giới thiệu cho bạn',
+        'mindSweep.group.funGrowth.p3': 'Sở thích đã bị tạm dừng',
+        'mindSweep.group.funGrowth.p4': 'Sự kiện hoặc vé cần đặt trước khi hết',
+        'mindSweep.group.funGrowth.p5': 'Điều gì đó bạn tò mò muốn học chỉ để vui',
+        'mindSweep.group.commitments.title': 'Lời hứa & dự án',
+        'mindSweep.group.commitments.p1': 'Bất cứ điều gì bạn nói với ai đó sẽ làm',
+        'mindSweep.group.commitments.p2': 'Dự án bạn đã cam kết nhưng chưa bắt đầu',
+        'mindSweep.group.commitments.p3': 'Công việc bạn đã bắt đầu và bỏ dở giữa chừng',
+        'mindSweep.group.commitments.p4': 'Quyết định bạn nợ ai đó',
+        'mindSweep.group.commitments.p5': 'Hạn chót gần hơn bạn cảm thấy',
+        'mindSweep.group.workComms.title': 'Con người & tin nhắn',
+        'mindSweep.group.workComms.p1': 'Email hoặc tin nhắn vẫn đang chờ phản hồi của bạn',
+        'mindSweep.group.workComms.p2': 'Cuộc họp bạn cần lên lịch, dời hoặc hủy',
+        'mindSweep.group.workComms.p3': 'Phản hồi bạn định đưa ra',
+        'mindSweep.group.workComms.p4': 'Điều gì đó bạn đang chờ từ người khác',
+        'mindSweep.group.workComms.p5': 'Giới thiệu hoặc bàn giao bạn đã nói sẽ thực hiện',
+        'mindSweep.group.workCraft.title': 'Ý tưởng & cải tiến',
+        'mindSweep.group.workCraft.p1': 'Phiền toái lặp đi lặp lại bạn có thể sửa một lần cho xong',
+        'mindSweep.group.workCraft.p2': 'Công cụ hoặc thiết lập cần sửa hoặc học',
+        'mindSweep.group.workCraft.p3': 'Ý tưởng bạn chưa viết xuống hoặc đề xuất',
+        'mindSweep.group.workCraft.p4': 'Tài liệu hoặc ghi chú bạn cứ hoãn lại',
+        'mindSweep.group.workCraft.p5': 'Việc bạn làm thủ công có thể tự động hóa',
+        'mindSweep.group.workAdmin.title': 'Hành chính công việc',
+        'mindSweep.group.workAdmin.p1': 'Chi phí, hóa đơn hoặc bảng chấm công đang chồng chất',
+        'mindSweep.group.workAdmin.p2': 'Tệp hoặc thư mục cần sắp xếp',
+        'mindSweep.group.workAdmin.p3': 'Đào tạo, đánh giá hoặc biểu mẫu có ngày đến hạn',
+        'mindSweep.group.workAdmin.p4': 'Đi lại, thiết bị hoặc quyền truy cập bạn cần thu xếp',
+        'mindSweep.group.workAdmin.p5': 'Bất cứ điều gì bản thân tương lai của bạn sẽ ước là đã lưu hôm nay',
+};
