+++ scripts/i18n-locale-parity.ts
@@ -27,6 +27,7 @@ const LOCALES: LocaleTarget[] = [
     { locale: 'pt', path: 'packages/core/src/i18n/locales/pt.ts' },
     { locale: 'ru', path: 'packages/core/src/i18n/locales/ru.ts' },
     { locale: 'tr', path: 'packages/core/src/i18n/locales/tr.ts' },
+    { locale: 'vi', path: 'packages/core/src/i18n/locales/vi.ts' },
     { locale: 'zh-Hans', path: 'packages/core/src/i18n/locales/zh-Hans.ts', fullParity: true },
     { locale: 'zh-Hant', path: 'packages/core/src/i18n/locales/zh-Hant.ts', fullParity: true },
 ];
